### 2025-09-26 #9082 @editorialbot
[PRE REVIEW]: ecodive: Parallel and Memory-Efficient R Package for Ecological Diversity Analysis
**Submitting author:** @dansmith01 (<a href="http://orcid.org/0000-0002-2479-2044">Daniel Patrick Smith</a>)
**Repository:** https://github.com/cmmr/ecodive
**Branch with paper.md** (empty if default branch):
**Version:** v2.2.0
