### 2025-09-26 65ef4eb
remove invalid doi
