+++ joss/paper.bib
@@ -172,8 +172,7 @@
   journal = {Scandinavian Journal of Statistics},
   volume  = {11},   
   pages   = {265-270},
-  url     = {https://www.jstor.org/stable/4616294},
-  doi     = {10.2307/4616294}
+  url     = {https://www.jstor.org/stable/4616294}
 }
 
 @article{Chao1992,
