+++ man/bdiv_functions.Rd
@@ -137,14 +137,13 @@ distance matrix to calculate. See examples.}
 default, \code{n_cpus()}, will use all logical CPU cores.}
 
 \item{norm}{Normalize the incoming counts. Options are:
-
-\if{html}{\out{<div class="sourceCode">}}\preformatted{   * `norm = 'percent'` - Relative abundance (sample abundances sum to 1).
-   * `norm = 'binary'`  - Unweighted presence/absence (each count is either 0 or 1).
-   * `norm = 'clr'`     - Centered log ratio.
-   * `norm = 'none'`    - No transformation.
-   
-   Default: `'percent'`, which is the expected input for these formulas.
-}\if{html}{\out{</div>}}}
+\describe{
+\item{\code{norm = "percent"} - }{ Relative abundance (sample abundances sum to 1). }
+\item{\code{norm = "binary"}  - }{ Unweighted presence/absence (each count is either 0 or 1). }
+\item{\code{norm = "clr"}     - }{ Centered log ratio. }
+\item{\code{norm = "none"}    - }{ No transformation. }
+}
+Default: \code{'percent'}, which is the expected input for these formulas.}
 
 \item{power}{Scaling factor for the magnitude of differences between
 communities (\eqn{p}). Default: \code{1.5}}
