+++ man/adiv_functions.Rd
@@ -59,14 +59,13 @@ Default: \code{10}.}
 default, \code{n_cpus()}, will use all logical CPU cores.}
 
 \item{norm}{Normalize the incoming counts. Options are:
-
-\if{html}{\out{<div class="sourceCode">}}\preformatted{   * `norm = 'percent'` - Relative abundance (sample abundances sum to 1).
-   * `norm = 'binary'`  - Unweighted presence/absence (each count is either 0 or 1).
-   * `norm = 'clr'`     - Centered log ratio.
-   * `norm = 'none'`    - No transformation.
-   
-   Default: `'percent'`, which is the expected input for these formulas.
-}\if{html}{\out{</div>}}}
+\describe{
+\item{\code{norm = "percent"} - }{ Relative abundance (sample abundances sum to 1). }
+\item{\code{norm = "binary"}  - }{ Unweighted presence/absence (each count is either 0 or 1). }
+\item{\code{norm = "clr"}     - }{ Centered log ratio. }
+\item{\code{norm = "none"}    - }{ No transformation. }
+}
+Default: \code{'percent'}, which is the expected input for these formulas.}
 
 \item{tree}{A \code{phylo}-class object representing the phylogenetic tree for
 the OTUs in \code{counts}. The OTU identifiers given by \code{colnames(counts)}
