+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   benchmark: benchmark.html
   ecodive: ecodive.html
   unifrac: unifrac.html
-last_built: 2025-09-25T18:28Z
+last_built: 2025-09-26T01:26Z
 urls:
   reference: https://cmmr.github.io/ecodive/reference
   article: https://cmmr.github.io/ecodive/articles
