+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   benchmark: benchmark.html
   ecodive: ecodive.html
   unifrac: unifrac.html
-last_built: 2025-09-26T01:26Z
+last_built: 2025-09-26T13:06Z
 urls:
   reference: https://cmmr.github.io/ecodive/reference
   article: https://cmmr.github.io/ecodive/articles
