### 2025-11-28 cb1bbd4
Operate on sparse matrices internally.
