+++ src/ecomatrix.c
@@ -1,0 +1,617 @@
+// Copyright (c) 2025 ecodive authors
+// Licensed under the MIT License: https://opensource.org/license/mit
+
+
+
+#include <R.h>
+#include <Rinternals.h>
+#include <string.h> // memcpy
+#include "get.h"
+#include "ecomatrix.h"
+#include "normalize.h"
+#include "memory.h"
+
+
+//=========================================================
+// Ensure we don't overwrite values in other R objects.
+//=========================================================
+
+static void rw_vec (void **ptr, int n_bytes) {
+  
+  if (is_safe_ptr(*ptr)) return;
+  
+  void *new_ptr = safe_malloc(n_bytes);
+  if (*ptr) memcpy(new_ptr, *ptr, n_bytes);
+  *ptr = new_ptr;
+}
+
+static int* rw_sam_vec (ecomatrix_t *em) {
+  rw_vec((void**)&(em->sam_vec), em->nnz * sizeof(int));
+  return em->sam_vec;
+}
+
+static int* rw_pos_vec (ecomatrix_t *em) {
+  int vec_len = em->n_samples + 1;
+  rw_vec((void**)&(em->pos_vec), vec_len * sizeof(int));
+  return em->pos_vec;
+}
+
+static int* rw_otu_vec (ecomatrix_t *em) {
+  rw_vec((void**)&(em->otu_vec), em->nnz * sizeof(int));
+  return em->otu_vec;
+}
+
+double* rw_val_vec (ecomatrix_t *em) {
+  rw_vec((void**)&(em->val_vec), em->nnz * sizeof(double));
+  return em->val_vec;
+}
+
+double* rw_clr_vec (ecomatrix_t *em) {
+  rw_vec((void**)&(em->clr_vec), em->n_samples * sizeof(double));
+  return em->clr_vec;
+}
+
+
+
+//=========================================================
+// Accept double, integer, or logical values.
+//=========================================================
+
+static void assign_sparse_vals(ecomatrix_t *em, SEXP sexp_vals) {
+  
+  if (isReal(sexp_vals)) {
+    em->val_vec = REAL(sexp_vals);
+  }
+  
+  else if (isInteger(sexp_vals) || isLogical(sexp_vals)) {
+    
+    double *val_vec = rw_val_vec(em);
+    
+    int  nnz = em->nnz;
+    int *tmp = INTEGER(sexp_vals);
+    for (int i = 0; i < nnz; i++) {
+      val_vec[i] = (double)(tmp[i]);
+    }
+  }
+  
+  else {
+    error("Input must be numeric");
+  }
+  
+}
+
+
+
+
+static void assign_dense_vals(ecomatrix_t *em, SEXP sexp_vals, int margin) {
+  
+  int n_samples = em->n_samples;
+  int n_otus    = em->n_otus;
+  int mtx_len   = length(sexp_vals);
+  
+  double *dbl_vals = NULL;
+  int    *int_vals = NULL;
+  
+  
+  // Accept double, integer, or logical values.
+  // --------------------------------------------
+  
+  if (isReal(sexp_vals)) {
+    dbl_vals = REAL(sexp_vals);
+  }
+  else if (isInteger(sexp_vals) || isLogical(sexp_vals)) {
+    int_vals = INTEGER(sexp_vals);
+  }
+  else {
+    error("Input must be numeric");
+  }
+  
+  
+  // Count the number of non-zero entries.
+  // --------------------------------------------
+  
+  int nnz = 0;
+  
+  if (dbl_vals != NULL) {
+    for (int i = 0; i < mtx_len; i++) {
+      if (dbl_vals[i]) nnz++;
+    }
+  }
+  else {
+    for (int i = 0; i < mtx_len; i++) {
[…]
