+++ src/beta_div.c
@@ -6,7 +6,10 @@
 #include <R.h>
 #include <Rinternals.h>
 #include <math.h>   // fabs, log, pow, sqrt
-#include <stdlib.h> // calloc, free
+#include <string.h> // memset
+#include "ecomatrix.h"
+#include "normalize.h"
+#include "memory.h"
 
 // Detect if pthread is available.
 #if defined __has_include
@@ -15,6 +18,8 @@
 #    define HAVE_PTHREAD
 #  endif
 #endif
+
+typedef void *(*pthread_func_t)(void *);
 
 #define BDIV_BHATTACHARYYA  1
 #define BDIV_BRAY           2
@@ -40,67 +45,160 @@
 #define BDIV_SQUARED_CHORD 23
 #define BDIV_WAVE_HEDGES   24
 
-static int     algorithm;
-static double *otu_mtx;
-static int     n_otus;
 static int     n_samples;
+static int     n_otus;
+static int     n_dist;
 static int     n_pairs;
 static int     n_threads;
+static int    *pos_vec;
+static int    *otu_vec;
+static double *val_vec;
+static double *clr_vec;
 static int    *pairs_vec;
-static char    all_pairs;
 static double *dist_vec;
 static SEXP   *sexp_extra;
 
 
 /*
- * The START_PAIR_LOOP and END_PAIR_LOOP macros efficiently 
- * loop through all combinations of samples. Skips unwanted 
- * pairings and pairings not assigned to the current thread. 
- * Ensures all threads process the same number of pairs.
+ * The FOREACH_PAIR macro iterates through all combinations of 
+ * samples. Skips unwanted pairings and pairings not assigned to 
+ * the current thread. Ensures all threads process the same 
+ * number of pairs. The code should assign to `distance`.
  * 
- * After calling START_PAIR_LOOP the code can expect `x1` 
- * and `y1` to be the two samples' first indices in 
- * `otu_mtx`. The code should assign to `distance` before 
- * calling END_PAIR_LOOP.
- * 
- * After calling START_OTU_LOOP the code can expect `x` 
- * and `y` to be the current OTU's values from `otu_mtx`.
+ * The FOREACH_OTU macro iterates through all OTU abundances for 
+ * a given pair of samples, assigning the values to `x` and `y`.
  * 
  * Implemented as macros to avoid the overhead of a function
  * call or the messiness of duplicated code.
  */
-
-#define START_PAIR_LOOP                                        \
-int thread_i = *((int *) arg);                                 \
-int pair_idx = 0; /* The pairs that we're asked to compute */  \
-int dist_idx = 0; /* All combinations; length of return vec */ \
-for (int x1 = 0; x1 < n_samples - 1; x1++) {                   \
-  for (int y1 = x1 + 1; y1 < n_samples; y1++) {                \
-    if (all_pairs || pairs_vec[pair_idx] == dist_idx) {        \
-      if (pair_idx % n_threads == thread_i) {                  \
-        double distance = 0;                                   \
-
-        
-#define START_OTU_LOOP                                         \
-        for (int otu = 0; otu < n_otus; otu++) {               \
-          double x = otu_mtx[x1 + otu * n_samples];            \
-          double y = otu_mtx[y1 + otu * n_samples];            \
-
-
-#define END_OTU_LOOP                                           \
+#define FOREACH_PAIR(expression)                               \
+  do {                                                         \
+    int thread_i = *((int *) arg);                             \
+    int dist_idx = 0;                                          \
+                                                               \
+    if (pairs_vec == NULL) { /* All vs All */                  \
+                                                               \
+      for (int sam_i = 0; sam_i < n_samples - 1; sam_i++) {    \
+        for (int sam_j = sam_i + 1; sam_j < n_samples; sam_j++) {\
[…]
