+++ articles/index.md
@@ -1,0 +1,13 @@
+# Articles
+
+### All vignettes
+
+- [Alpha Diversity](https://cmmr.github.io/ecodive/articles/adiv.md):
+- [Selecting a Beta Diversity
+  Metric](https://cmmr.github.io/ecodive/articles/bdiv_guide.md):
+- [Beta Diversity](https://cmmr.github.io/ecodive/articles/bdiv.md):
+- [Benchmarks](https://cmmr.github.io/ecodive/articles/benchmark.md):
+- [Introduction to
+  ecodive](https://cmmr.github.io/ecodive/articles/ecodive.md):
+- [UniFrac
+  Calculations](https://cmmr.github.io/ecodive/articles/unifrac.md):
