+++ authors.md
@@ -1,0 +1,26 @@
+# Authors and Citation
+
+## Authors
+
+- **Daniel P. Smith**. Author, maintainer.
+  [](https://orcid.org/0000-0002-2479-2044)
+
+- **Alkek Center for Metagenomics and Microbiome Research**. Copyright
+  holder, funder.
+
+## Citation
+
+Source:
+[`DESCRIPTION`](https://github.com/cmmr/ecodive/blob/main/DESCRIPTION)
+
+Smith D (2025). *ecodive: Parallel and Memory-Efficient Ecological
+Diversity Metrics*. R package version 2.2.0,
+<https://cmmr.github.io/ecodive/>.
+
+    @Manual{,
+      title = {ecodive: Parallel and Memory-Efficient Ecological Diversity Metrics},
+      author = {Daniel P. Smith},
+      year = {2025},
+      note = {R package version 2.2.0},
+      url = {https://cmmr.github.io/ecodive/},
+    }
