+++ reference/ex_tree.md
@@ -1,0 +1,29 @@
+# Example phylogenetic tree
+
+Companion tree for `ex_counts`.
+
+## Usage
+
+``` r
+ex_tree
+```
+
+## Format
+
+A `phylo` object.
+
+## Details
+
+`ex_tree` encodes this tree structure:
+
+          +----------44---------- Haemophilus
+      +-2-|
+      |   +----------------68---------------- Bacteroides
+      |
+      |             +---18---- Streptococcus
+      |      +--12--|
+      |      |      +--11-- Staphylococcus
+      +--11--|
+             |      +-----24----- Corynebacterium
+             +--12--|
+                    +--13-- Propionibacterium
