+++ reference/ex_counts.md
@@ -1,0 +1,18 @@
+# Example counts matrix
+
+Genera found on four human body sites.
+
+## Usage
+
+``` r
+ex_counts
+```
+
+## Format
+
+A matrix of 4 samples (columns) x 6 genera (rows).
+
+## Source
+
+Derived from The Human Microbiome Project dataset.
+<https://commonfund.nih.gov/hmp>

[…]
