### 2025-11-28 e4eb26f
`ecomatrix.c` bug fixes and test coverage
