+++ tests/testthat/test-ecomatrix.r
@@ -1,0 +1,125 @@
+test_that("ecomatrix.c parsing logic is covered", {
+
+  # Ensure Matrix and slam packages are available for testing
+  skip_if_not_installed('Matrix')
+  skip_if_not_installed('slam')
+
+  # Define a test function that calls an internal C function
+  # We will test an alpha diversity metric which uses the ecomatrix parsing
+  test_parsing <- function(m, margin = 1L) {
+    observed(m, margin = margin)
+  }
+
+  # Expected result: number of non-zero entries per sample
+  expected_margin1 <- c(A = 3, B = 3, C = 3, D = 2)
+  expected_margin2 <- c(OTU1 = 2, OTU2 = 4, OTU3 = 4, OTU4 = 3, OTU5 = 2)
+
+  # === Test with base R matrix ===
+  m_base <- counts
+  expect_equal(test_parsing(m_base, margin = 1L), expected_margin1)
+  expect_equal(test_parsing(t(m_base), margin = 2L), expected_margin1)
+
+  # Test with logical matrix
+  m_logical <- m_base > 0
+  expect_equal(test_parsing(m_logical, margin = 1L), expected_margin1)
+  expect_equal(test_parsing(t(m_logical), margin = 2L), expected_margin1)
+
+  # Test with double matrix
+  m_double <- m_base * 1.1
+  expect_equal(test_parsing(m_double, margin = 1L), expected_margin1)
+  expect_equal(test_parsing(t(m_double), margin = 2L), expected_margin1)
+
+
+  # === Test with Matrix package types ===
+  m_base_t <- t(m_base)
+  m_dgC    <- as(m_base, "dgCMatrix")
+  m_dgC_t  <- as(m_base_t, "dgCMatrix")
+  m_dgT    <- as(m_base, "TsparseMatrix")
+  m_dgT_t  <- as(m_base_t, "TsparseMatrix")
+  m_dge    <- as(m_base, "unpackedMatrix")
+  m_dge_t  <- as(m_base_t, "unpackedMatrix")
+
+  # dgCMatrix
+  expect_equal(test_parsing(m_dgC, margin = 1L), expected_margin1, info = "dgCMatrix margin 1")
+  expect_equal(test_parsing(m_dgC_t, margin = 2L), expected_margin1, info = "dgCMatrix margin 2")
+
+  # dgTMatrix
+  expect_equal(test_parsing(m_dgT, margin = 1L), expected_margin1, info = "dgTMatrix margin 1")
+  expect_equal(test_parsing(m_dgT_t, margin = 2L), expected_margin1, info = "dgTMatrix margin 2")
+
+  # dgeMatrix
+  expect_equal(test_parsing(m_dge, margin = 1L), expected_margin1, info = "dgeMatrix margin 1")
+  expect_equal(test_parsing(m_dge_t, margin = 2L), expected_margin1, info = "dgeMatrix margin 2")
+
+  # Test with logical Matrix types
+  m_logical_t <- t(m_logical)
+  m_lgC       <- as(m_logical, "dgCMatrix")
+  m_lgC_t     <- as(m_logical_t, "dgCMatrix")
+  m_lgT       <- as(m_logical, "TsparseMatrix")
+  m_lge       <- as(m_logical, "unpackedMatrix")
+  expect_equal(test_parsing(m_lgC, margin = 1L), expected_margin1, info = "lgCMatrix margin 1")
+  expect_equal(test_parsing(m_lgC_t, margin = 2L), expected_margin1, info = "lgCMatrix margin 2")
+  expect_equal(test_parsing(m_lgT, margin = 1L), expected_margin1, info = "lgTMatrix margin 1")
[…]
