+++ .github/workflows/R-CMD-check.yaml
@@ -37,8 +37,6 @@ jobs:
           - {os: ubuntu-latest,    r: 'release'}
           - {os: ubuntu-latest,    r: 'oldrel-1'}
           - {os: ubuntu-22.04,     r: 'release'}
-          - {os: ubuntu-22.04-arm, r: 'release' }
-          - {os: ubuntu-24.04-arm, r: 'release' }
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}

[…]
