+++ .github/workflows/R-CMD-check.yaml
@@ -62,3 +62,8 @@ jobs:
         with:
           upload-snapshots: true
           build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf")'
+
+      - name: Show install log
+        if: always()
+        run: find . -name '00install.out' -exec cat {} +
+        shell: bash
