### 2025-11-28 c5dc9d1
omit `skip_on_cran()` from tests
