### 2025-12-01 5b93a97
valgrind fixes
