### 2025-12-01 4669e4f
more valgrind fixes
