+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: ecodive
 Type: Package
 Title: Parallel and Memory-Efficient Ecological Diversity Metrics
-Version: 2.2.0
+Version: 2.2.1
 Authors@R: c(
     person(
       "Daniel P.", "Smith",
