### 2025-12-01 ff2f94b
triplet bug; test coverage
