+++ src/ecomatrix.h
@@ -25,9 +25,6 @@ typedef struct {
 // ecomatrix constructor
 ecomatrix_t* new_ecomatrix(SEXP sexp_matrix, SEXP sexp_margin);
 
-// Quickly get the number of samples.
-int count_samples(SEXP mtx, SEXP sexp_margin);
-
 // avoid overwriting R object data
 double* rw_val_vec(ecomatrix_t *em);
 double* rw_clr_vec(ecomatrix_t *em);
