### 2025-12-01 f82534d
News update
