### 2025-12-01 3d3b229
figures, `void`s, `UNPROTECT`s
