+++ src/memory.h
@@ -8,7 +8,7 @@
 void  init_n_ptrs(int n);
 void* safe_malloc(size_t bytes);
 int   is_safe_ptr(void *ptr);
-void  free_all();
+void  free_all(void);
 void* free_one(void *ptr);
 void* maybe_free_one(void *ptr);
 
