+++ news/index.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## ecodive 2.2.0
+
+- Dedicated rarefaction backends for dense and sparse matrices.
+- Auto-conversion to col-major sparse matrix for distance/diversity
+  calculations.
+
 ## ecodive 2.1.0
 
 CRAN release: 2025-09-26
