### 2025-12-01 eca3f69
ubsan; yaml for figures
