+++ articles/benchmark.md
@@ -22,7 +22,7 @@ implementations.
 | [entropart](https://doi.org/10.32614/CRAN.package.entropart)       | ***Serial R***       | *none*                  |
 | [GUniFrac](https://doi.org/10.32614/CRAN.package.GUniFrac)         | *none*               | ***Serial C***          |
 | [labdsv](https://doi.org/10.32614/CRAN.package.labdsv)             | ***Serial FORTRAN*** | *none*                  |
-| [parallelDist](https://doi.org/10.32614/CRAN.package.parallelDist) | *none*               | ***Parallel C++***      |
+| [parallelDist](https://doi.org/10.32614/CRAN.package.parallelDist) | ***Parallel C++***   | *none*                  |
 | [philentropy](https://doi.org/10.32614/CRAN.package.philentropy)   | ***Serial C++***     | *none*                  |
 | [phyloregion](https://doi.org/10.32614/CRAN.package.phyloregion)   | vegan                | ***Serial R***          |
 | [phyloseq](https://doi.org/doi:10.18129/B9.bioc.phyloseq)          | vegan                | ***Parallel R***        |
@@ -39,7 +39,7 @@ the following hardware and software configuration:
 
     CPU: 6-Core Intel i5-9600K @ 3.70GHz
     RAM: 64.0 GB
-    OS: Windows 11 Pro (64-bit, Version 24H2, Build 26100.4652)
+    OS: Windows 11 Pro (64-bit, Version 25H2, Build 26200.7171)
 
 Furthermore, the `bench::mark()` function was utilized to verify that
 the outputs from all benchmarked expressions were numerically
@@ -82,39 +82,40 @@ pak::pkg_install(pkg = c(
 # Software Versions
 
 version$version.string
-#> [1] "R version 4.5.1 (2025-06-13 ucrt)"
+#> [1] "R version 4.5.2 (2025-10-31 ucrt)"
 
 data.frame(ver = sapply(FUN = packageDescription, fields = 'Version', c(
   'bench', 'dplyr', 'ggplot2', 'ggrepel', 'rbiom',
   'abdiv', 'adiv', 'ecodist', 'ecodive', 'entropart', 'GUniFrac', 
   'ampvis2', 'labdsv', 'parallelDist', 'philentropy', 
   'phyloregion', 'phyloseq', 'picante', 'tabula', 'vegan' )))
-#>                     ver
-#> bench             1.1.4
-#> dplyr             1.1.4
-#> ggplot2           3.5.2
-#> ggrepel           0.9.6
-#> rbiom             2.2.1
-#> abdiv             0.2.0
-#> adiv              2.2.1
-#> ecodist           2.1.3
-#> ecodive           2.0.0
-#> entropart        1.6-16
-#> GUniFrac            1.8
-#> ampvis2           2.8.9
-#> labdsv            2.1-2
-#> parallelDist      0.2.6
-#> philentropy       0.9.0
-#> phyloregion       1.0.9
-#> phyloseq         1.52.0
-#> picante           1.8.2
-#> tabula            3.3.1
-#> vegan             2.7-1
+#>                 ver
+#> bench         1.1.4
+#> dplyr         1.1.4
+#> ggplot2       4.0.1
+#> ggrepel       0.9.6
+#> rbiom         2.2.1
[…]
