### 2025-12-02 bacd927
benchmarks yaml
