+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   articles/benchmark: benchmark.html
   ecodive: ecodive.html
   articles/unifrac: unifrac.html
-last_built: 2025-12-02T20:06Z
+last_built: 2025-12-02T20:13Z
 urls:
   reference: https://cmmr.github.io/ecodive/reference
   article: https://cmmr.github.io/ecodive/articles
