### 2025-12-02 9eda589
Added Performance Guide Vignette
