+++ man/adiv_functions.Rd
@@ -50,7 +50,9 @@ squares(counts, margin = 1L, cpus = n_cpus())
 \item{counts}{A numeric matrix of count data where each column is a
 feature, and each row is a sample. Any object coercible with
 \code{as.matrix()} can be given here, as well as \code{phyloseq}, \code{rbiom},
-\code{SummarizedExperiment}, and \code{TreeSummarizedExperiment} objects.}
+\code{SummarizedExperiment}, and \code{TreeSummarizedExperiment} objects. For
+optimal performance with very large datasets, see the guide in
+\code{vignette('performance')}.}
 
 \item{cutoff}{The maximum number of observations to consider "rare".
 Default: \code{10}.}
