+++ llms.txt
@@ -283,5 +283,7 @@ By contributing to this project, you agree to abide by its terms.
 - [Benchmarks](https://cmmr.github.io/ecodive/articles/benchmark.md):
 - [Introduction to
   ecodive](https://cmmr.github.io/ecodive/articles/ecodive.md):
+- [Performance
+  Guide](https://cmmr.github.io/ecodive/articles/performance.md):
 - [UniFrac
   Calculations](https://cmmr.github.io/ecodive/articles/unifrac.md):
