### 2025-12-02 8d82ac4
updated relative speed ups
