+++ pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   ecodive: ecodive.html
   performance: performance.html
   articles/unifrac: unifrac.html
-last_built: 2025-12-02T21:15Z
+last_built: 2025-12-02T21:23Z
 urls:
   reference: https://cmmr.github.io/ecodive/reference
   article: https://cmmr.github.io/ecodive/articles
