+++ .github/workflows/R-CMD-check.yaml
@@ -32,16 +32,19 @@ jobs:
           - {os: macos-latest,     r: 'devel'}
           - {os: macos-latest,     r: 'release'}
           - {os: macos-latest,     r: 'oldrel-1'}
+          - {os: macos-latest,     r: '3.6'}
           - {os: macos-14,         r: 'release'}
           
           - {os: windows-latest,   r: 'devel'}
           - {os: windows-latest,   r: 'release'}
           - {os: windows-latest,   r: 'oldrel-1'}
+          - {os: windows-latest,   r: '3.6'}
           - {os: windows-2022,     r: 'release'}
           
           - {os: ubuntu-latest,    r: 'devel' }
           - {os: ubuntu-latest,    r: 'release'}
           - {os: ubuntu-latest,    r: 'oldrel-1'}
+          - {os: ubuntu-latest,    r: '3.6'}
           - {os: ubuntu-22.04,     r: 'release'}
 
     env:
