### 2025-12-02 d66e76d
support earlier `Matrix` versions in rarefy tests
