+++ .github/workflows/R-CMD-check.yaml
@@ -32,7 +32,6 @@ jobs:
           - {os: macos-latest,     r: 'devel'}
           - {os: macos-latest,     r: 'release'}
           - {os: macos-latest,     r: 'oldrel-1'}
-          - {os: macos-latest,     r: '3.6'}
           - {os: macos-14,         r: 'release'}
           
           - {os: windows-latest,   r: 'devel'}
