### 2025-12-02 ba3d763
support earlier `Matrix` versions in tests
