### 2025-12-08 1cb7f89
change unprotected `mkString` to gc protected `install`
