### 2025-12-08 56cda82
Custom `rchk` workflow
