### 2025-12-08 a585b27
rchk using `--no-manual --no-build-vignettes`
