### 2025-12-08 41cfd2e
test `rchk`
