### 2025-12-08 873ee5b
show install log for rchk
