### 2025-12-08 011aa52
print rchk errors
