### 2025-12-08 c4fa1f3
rchk testing complete
