### 2025-12-08 f05d357
corrected CRAN `rchk` issue
