+++ .github/workflows/rhub.yaml
@@ -76,35 +76,60 @@ jobs:
           # Define paths used by the r-hub rchk container
           pkg           <- read.dcf("DESCRIPTION", "Package")[[1]]
           libs_dir      <- file.path("/opt/R/devel-rchk/packages/lib", pkg, "libs")
+          so_file       <- file.path(libs_dir, paste0(pkg, ".so"))
           bcheck_file   <- file.path(libs_dir, paste0(pkg, ".so.bcheck"))
           maacheck_file <- file.path(libs_dir, paste0(pkg, ".so.maacheck"))
 
-          errors <- character(0)
-
-          # 1. Check bcheck (using custom logic)
-          cat("Looking for bcheck file at ", bcheck_file, "\n")
-          if (file.exists(bcheck_file)) {
-            cat("Reading bcheck file\n")
-            b_lines <- readLines(bcheck_file)
-            b_lines <- b_lines[!grepl("too many states", b_lines)]
-            b_lines <- b_lines[!grepl("^Analyzed [0-9]+ functions", b_lines)]
-            errors <- c(errors, b_lines)
+          # rchk.sh often skips maacheck, so we run it manually here.
+          if (file.exists(so_file)) {
+            cat("Running maacheck manually on", so_file, "\n")
+            # maacheck is in the PATH in this container
+            cmd <- paste("maacheck", shQuote(so_file), ">", shQuote(maacheck_file), "2>&1")
+            exit_code <- system(cmd)
+            if (exit_code != 0) cat("Warning: maacheck command returned non-zero exit code\n")
+          } else {
+            stop("Could not find shared object file at: ", so_file)
           }
 
-          # 2. Check maacheck (missing in standard rhub check)
-          cat("Looking for maacheck file at ", maacheck_file, "\n")
+          errors <- character(0)
+
+          # 1. Check bcheck (generated by the previous run-check step)
+          cat("Checking bcheck file:", bcheck_file, "\n")
+          if (file.exists(bcheck_file)) {
+            b_lines <- readLines(bcheck_file)
+            # Filter out known non-error lines from bcheck
+            b_lines <- b_lines[!grepl("too many states", b_lines)]
+            b_lines <- b_lines[!grepl("^Analyzed [0-9]+ functions", b_lines)]
+            b_lines <- b_lines[!grepl("^==== rchk bcheck", b_lines)]
+            if (length(b_lines) > 0) {
+               cat("Found bcheck errors:\n")
+               print(b_lines)
+            }
+            errors <- c(errors, b_lines)
+          } else {
+            cat("WARNING: bcheck file not found!\n")
+          }
+
+          # 2. Check maacheck (generated manually above)
+          cat("Checking maacheck file:", maacheck_file, "\n")
           if (file.exists(maacheck_file)) {
-            cat("Reading maacheck file\n")
             m_lines <- readLines(maacheck_file)
-            errors  <- c(errors, m_lines)
+            # Filter out standard maacheck headers if necessary, though usually it's just errors/warnings
+            if (length(m_lines) > 0) {
+               cat("Found maacheck output:\n")
+               print(m_lines)
+               errors <- c(errors, m_lines)
+            }
+          } else {
+            cat("WARNING: maacheck file not found!\n")
           }
 
           # 3. Fail if errors exist
           errors <- errors[nzchar(errors)]
           if (length(errors) > 0) {
-            stop("rchk failed (enhanced check)! Errors detected:\n", paste(errors, collapse = "\n"))
+            stop("rchk failed! Errors detected:\n", paste(errors, collapse = "\n"))
           } else {
-            cat("SUCCESS: No memory protection errors found by enhanced rchk.\n")
+            cat("SUCCESS: No memory protection errors found by rchk.\n")
           }
 
       - name: Show install log
