+++ .github/workflows/rhub.yaml
@@ -80,13 +80,28 @@ jobs:
           bcheck_file   <- file.path(libs_dir, paste0(pkg, ".so.bcheck"))
           maacheck_file <- file.path(libs_dir, paste0(pkg, ".so.maacheck"))
 
-          # rchk.sh often skips maacheck, so we run it manually here.
           if (file.exists(so_file)) {
-            cat("Running maacheck manually on", so_file, "\n")
-            # maacheck is in the PATH in this container
-            cmd <- paste("maacheck", shQuote(so_file), ">", shQuote(maacheck_file), "2>&1")
-            exit_code <- system(cmd)
-            if (exit_code != 0) cat("Warning: maacheck command returned non-zero exit code\n")
+            # Locate maacheck (it is often in /rchk/scripts or /opt/rchk)
+            maacheck_bin <- Sys.which("maacheck")
+            if (maacheck_bin == "") {
+                cat("maacheck not in PATH. Searching...\n")
+                # Search common locations or root to find it
+                found <- system("find /rchk /opt -name maacheck -type f -executable 2>/dev/null", intern = TRUE)
+                if (length(found) > 0) maacheck_bin <- found[1]
+            }
+
+            if (length(maacheck_bin) > 0 && maacheck_bin != "") {
+                cat("Found maacheck at:", maacheck_bin, "\n")
+                cat("Running maacheck on", so_file, "\n")
+                
+                # Run maacheck and redirect output
+                cmd <- paste(maacheck_bin, shQuote(so_file), ">", shQuote(maacheck_file), "2>&1")
+                exit_code <- system(cmd)
+                
+                if (exit_code != 0) cat("Warning: maacheck command returned non-zero exit code\n")
+            } else {
+                cat("Error: Could not find 'maacheck' executable in container.\n")
+            }
           } else {
             stop("Could not find shared object file at: ", so_file)
           }
@@ -97,7 +112,7 @@ jobs:
           cat("Checking bcheck file:", bcheck_file, "\n")
           if (file.exists(bcheck_file)) {
             b_lines <- readLines(bcheck_file)
-            # Filter out known non-error lines from bcheck
+            # Filter out known non-error lines
             b_lines <- b_lines[!grepl("too many states", b_lines)]
             b_lines <- b_lines[!grepl("^Analyzed [0-9]+ functions", b_lines)]
             b_lines <- b_lines[!grepl("^==== rchk bcheck", b_lines)]
@@ -107,21 +122,20 @@ jobs:
             }
             errors <- c(errors, b_lines)
           } else {
-            cat("WARNING: bcheck file not found!\n")
+            cat("WARNING: bcheck file not found! (Did rchk.sh run successfully?)\n")
           }
 
           # 2. Check maacheck (generated manually above)
           cat("Checking maacheck file:", maacheck_file, "\n")
           if (file.exists(maacheck_file)) {
             m_lines <- readLines(maacheck_file)
-            # Filter out standard maacheck headers if necessary, though usually it's just errors/warnings
             if (length(m_lines) > 0) {
                cat("Found maacheck output:\n")
                print(m_lines)
                errors <- c(errors, m_lines)
             }
           } else {
-            cat("WARNING: maacheck file not found!\n")
+             cat("WARNING: maacheck file was not generated.\n")
           }
 
           # 3. Fail if errors exist
