### 2025-12-09 b52e7d3
run `extract-bc` before `maacheck`
