+++ .github/workflows/rhub.yaml
@@ -73,72 +73,92 @@ jobs:
         if: ${{ always() && matrix.config.label == 'rchk' }}
         shell: Rscript {0}
         run: |
-          # Define paths used by the r-hub rchk container
+          # Define paths
           pkg           <- read.dcf("DESCRIPTION", "Package")[[1]]
           libs_dir      <- file.path("/opt/R/devel-rchk/packages/lib", pkg, "libs")
           so_file       <- file.path(libs_dir, paste0(pkg, ".so"))
+          bc_file       <- file.path(libs_dir, paste0(pkg, ".so.bc"))
           bcheck_file   <- file.path(libs_dir, paste0(pkg, ".so.bcheck"))
           maacheck_file <- file.path(libs_dir, paste0(pkg, ".so.maacheck"))
 
-          if (file.exists(so_file)) {
-            # Locate maacheck (it is often in /rchk/scripts or /opt/rchk)
-            maacheck_bin <- Sys.which("maacheck")
-            if (maacheck_bin == "") {
-                cat("maacheck not in PATH. Searching...\n")
-                # Search common locations or root to find it
-                found <- system("find /rchk /opt -name maacheck -type f -executable 2>/dev/null", intern = TRUE)
-                if (length(found) > 0) maacheck_bin <- found[1]
+          # --- HELPER: Find executable ---
+          find_exec <- function(name) {
+            path <- Sys.which(name)
+            if (path == "") {
+              # Common locations in rchk images
+              candidates <- system(paste("find /rchk /opt /usr -name", name, "-type f -executable 2>/dev/null"), intern = TRUE)
+              if (length(candidates) > 0) path <- candidates[1]
             }
+            return(path)
+          }
 
-            if (length(maacheck_bin) > 0 && maacheck_bin != "") {
-                cat("Found maacheck at:", maacheck_bin, "\n")
-                cat("Running maacheck on", so_file, "\n")
-                
-                # Run maacheck and redirect output
-                cmd <- paste(maacheck_bin, shQuote(so_file), ">", shQuote(maacheck_file), "2>&1")
-                exit_code <- system(cmd)
-                
-                if (exit_code != 0) cat("Warning: maacheck command returned non-zero exit code\n")
-            } else {
-                cat("Error: Could not find 'maacheck' executable in container.\n")
+          # --- 1. GENERATE BITCODE (.bc) ---
+          if (file.exists(so_file)) {
+            if (!file.exists(bc_file)) {
+              cat("Bitcode file not found. Attempting to extract from .so...\n")
+              extract_bc <- find_exec("extract-bc")
+              
+              if (length(extract_bc) > 0 && extract_bc != "") {
+                cmd <- paste(extract_bc, shQuote(so_file))
+                cat("Running:", cmd, "\n")
+                if (system(cmd) == 0) {
+                  cat("Successfully extracted bitcode to:", bc_file, "\n")
+                } else {
+                  cat("Error: extract-bc failed.\n")
+                }
+              } else {
+                cat("Error: 'extract-bc' tool not found.\n")
+              }
             }
           } else {
             stop("Could not find shared object file at: ", so_file)
           }
 
+          # --- 2. RUN MAACHECK ---
+          if (file.exists(bc_file)) {
+            maacheck_bin <- find_exec("maacheck")
+            
+            if (length(maacheck_bin) > 0 && maacheck_bin != "") {
+                cat("Running maacheck on", bc_file, "\n")
+                # Run maacheck on the .bc file
+                cmd <- paste(maacheck_bin, shQuote(bc_file), ">", shQuote(maacheck_file), "2>&1")
+                system(cmd)
+            } else {
+                cat("Error: Could not find 'maacheck' executable.\n")
+            }
[…]
