### 2025-12-09 ff5c9e2
extract libR.so bitcode too
