+++ .github/workflows/rhub.yaml
@@ -73,64 +73,70 @@ jobs:
         if: ${{ always() && matrix.config.label == 'rchk' }}
         shell: Rscript {0}
         run: |
-          # Define paths
+          # --- CONFIGURATION ---
           pkg           <- read.dcf("DESCRIPTION", "Package")[[1]]
           libs_dir      <- file.path("/opt/R/devel-rchk/packages/lib", pkg, "libs")
           so_file       <- file.path(libs_dir, paste0(pkg, ".so"))
-          bc_file       <- file.path(libs_dir, paste0(pkg, ".so.bc"))
+          pkg_bc_file   <- file.path(libs_dir, paste0(pkg, ".so.bc"))
           bcheck_file   <- file.path(libs_dir, paste0(pkg, ".so.bcheck"))
           maacheck_file <- file.path(libs_dir, paste0(pkg, ".so.maacheck"))
+          
+          # Path to R's shared library (needed for maacheck context)
+          r_lib_file    <- file.path(R.home("lib"), "libR.so")
+          r_bc_file     <- "libR.so.bc" # We will create this in the working dir
 
           # --- HELPER: Find executable ---
           find_exec <- function(name) {
             path <- Sys.which(name)
             if (path == "") {
-              # Common locations in rchk images
               candidates <- system(paste("find /rchk /opt /usr -name", name, "-type f -executable 2>/dev/null"), intern = TRUE)
               if (length(candidates) > 0) path <- candidates[1]
             }
             return(path)
           }
 
-          # --- 1. GENERATE BITCODE (.bc) ---
-          if (file.exists(so_file)) {
-            if (!file.exists(bc_file)) {
-              cat("Bitcode file not found. Attempting to extract from .so...\n")
-              extract_bc <- find_exec("extract-bc")
-              
-              if (length(extract_bc) > 0 && extract_bc != "") {
-                cmd <- paste(extract_bc, shQuote(so_file))
-                cat("Running:", cmd, "\n")
-                if (system(cmd) == 0) {
-                  cat("Successfully extracted bitcode to:", bc_file, "\n")
-                } else {
-                  cat("Error: extract-bc failed.\n")
-                }
-              } else {
-                cat("Error: 'extract-bc' tool not found.\n")
-              }
-            }
-          } else {
-            stop("Could not find shared object file at: ", so_file)
+          extract_bc <- find_exec("extract-bc")
+          maacheck   <- find_exec("maacheck")
+
+          # --- 1. PREPARE R BITCODE ---
+          # We copy libR.so to the current dir to avoid permission issues in /opt
+          if (file.exists(r_lib_file)) {
+             file.copy(r_lib_file, "libR.so", overwrite = TRUE)
+             if (file.exists("libR.so") && length(extract_bc) > 0) {
+                cat("Extracting bitcode from libR.so...\n")
+                system(paste(extract_bc, "libR.so"))
+             }
+          }
+          
+          if (!file.exists(r_bc_file)) {
+             cat("WARNING: Could not generate libR.so.bc. maacheck will likely fail.\n")
           }
 
-          # --- 2. RUN MAACHECK ---
-          if (file.exists(bc_file)) {
-            maacheck_bin <- find_exec("maacheck")
-            
-            if (length(maacheck_bin) > 0 && maacheck_bin != "") {
-                cat("Running maacheck on", bc_file, "\n")
-                # Run maacheck on the .bc file
-                cmd <- paste(maacheck_bin, shQuote(bc_file), ">", shQuote(maacheck_file), "2>&1")
-                system(cmd)
-            } else {
-                cat("Error: Could not find 'maacheck' executable.\n")
-            }
-          } else {
[…]
