### 2025-12-09 1b43b5c
rely on custom rchk
