+++ .github/workflows/rhub.yaml
@@ -69,109 +69,6 @@ jobs:
           token: ${{ secrets.RHUB_TOKEN }}
           job-config: ${{ matrix.config.job-config }}
 
-      - name: Verify rchk results (Enhanced)
-        if: ${{ always() && matrix.config.label == 'rchk' }}
-        shell: Rscript {0}
-        run: |
-          # --- CONFIGURATION ---
-          pkg           <- read.dcf("DESCRIPTION", "Package")[[1]]
-          libs_dir      <- file.path("/opt/R/devel-rchk/packages/lib", pkg, "libs")
-          so_file       <- file.path(libs_dir, paste0(pkg, ".so"))
-          pkg_bc_file   <- file.path(libs_dir, paste0(pkg, ".so.bc"))
-          bcheck_file   <- file.path(libs_dir, paste0(pkg, ".so.bcheck"))
-          maacheck_file <- file.path(libs_dir, paste0(pkg, ".so.maacheck"))
-          
-          # Path to R's shared library (needed for maacheck context)
-          r_lib_file    <- file.path(R.home("lib"), "libR.so")
-          r_bc_file     <- "libR.so.bc" # We will create this in the working dir
-
-          # --- HELPER: Find executable ---
-          find_exec <- function(name) {
-            path <- Sys.which(name)
-            if (path == "") {
-              candidates <- system(paste("find /rchk /opt /usr -name", name, "-type f -executable 2>/dev/null"), intern = TRUE)
-              if (length(candidates) > 0) path <- candidates[1]
-            }
-            return(path)
-          }
-
-          extract_bc <- find_exec("extract-bc")
-          maacheck   <- find_exec("maacheck")
-
-          # --- 1. PREPARE R BITCODE ---
-          # We copy libR.so to the current dir to avoid permission issues in /opt
-          if (file.exists(r_lib_file)) {
-             file.copy(r_lib_file, "libR.so", overwrite = TRUE)
-             if (file.exists("libR.so") && length(extract_bc) > 0) {
-                cat("Extracting bitcode from libR.so...\n")
-                system(paste(extract_bc, "libR.so"))
-             }
-          }
-          
-          if (!file.exists(r_bc_file)) {
-             cat("WARNING: Could not generate libR.so.bc. maacheck will likely fail.\n")
-          }
-
-          # --- 2. PREPARE PACKAGE BITCODE ---
-          if (file.exists(so_file) && length(extract_bc) > 0) {
-             if (!file.exists(pkg_bc_file)) {
-                cat("Extracting bitcode from package .so...\n")
-                system(paste(extract_bc, shQuote(so_file)))
-             }
-          }
-
-          # --- 3. RUN MAACHECK ---
-          if (file.exists(pkg_bc_file) && length(maacheck) > 0) {
-             # ARGUMENT ORDER MATTERS: maacheck <R_BITCODE> <PACKAGE_BITCODE>
-             # If R bitcode is missing, we try running without it (will fail with R_gc_internal error, but consistent)
-             args <- shQuote(pkg_bc_file)
-             if (file.exists(r_bc_file)) {
-                args <- paste(shQuote(r_bc_file), args)
-             }
-             
-             cmd <- paste(maacheck, args, ">", shQuote(maacheck_file), "2>&1")
-             cat("Running:", cmd, "\n")
-             system(cmd)
-          } else {
-             cat("Skipping maacheck: Missing tools or bitcode files.\n")
-          }
-
-          # --- 4. VERIFY RESULTS ---
-          errors <- character(0)
-
-          # Check bcheck
-          if (file.exists(bcheck_file)) {
-            b_lines <- readLines(bcheck_file)
-            b_lines <- b_lines[!grepl("too many states", b_lines)]
[…]
