+++ .github/workflows/alpine-musl-check.yaml
@@ -1,0 +1,23 @@
+name: Alpine Musl Check
+
+on:
+  workflow_dispatch:
+  push:
+    branches: [main]
+    paths:
+      - R/**
+      - src/**
+      - tests/**
+      - .Rbuildignore
+      - .github/workflows/alpine-musl-check.yaml
+
+jobs:
+  alpine-check:
+    runs-on: ubuntu-latest
+    
+    steps:
+      - name: Checkout Repository
+        uses: actions/checkout@v4
+
+      - name: Run Alpine Musl Check
+        uses: ./.github/actions/alpine-musl-check
