+++ .github/actions/alpine-musl-check/action.yaml
@@ -43,7 +43,7 @@ runs:
 
             echo "--- Running R CMD check ---"
             # Capture exit status instead of failing immediately (due to set -e)
-            if R -q -e "rcmdcheck::rcmdcheck(args = c(\"--no-manual\", \"--as-cran\", \"--no-vignettes\"))"; then
+            if R -q -e "rcmdcheck::rcmdcheck(args = c(\"--as-cran\", \"--no-manual\", \"--no-vignettes\"), build_args = c(\"--no-manual\", \"no-build-vignettes\"))"; then
               STATUS=0
               echo "✅ Check Succeeded"
             else
