### 2025-12-11 03115df
provide `.registration = TRUE` and pandoc for musl check
