### 2025-12-19 #9082 @goldingn
Hi @CosteaPaul, @kelly-sovacool, @sganesh5, @qpmnguyen would any of you be interested in reviewing this submission on ecological(/microbiome) diversity metrics for JOSS?
