### 2025-12-19 1dfed2d
updated paths for JOSS draft PDF.
