### 2026-01-15 5b2df77
Drop explicit zeroes from sparse matrices after rarefaction.
