### 2026-01-15 4cb88b8
Addresses `rchk`'s GC paranoia.
