+++ src/rarefy.c
@@ -1,24 +1,8 @@
 // Copyright (c) 2025 ecodive authors
 // Licensed under the MIT License: https://opensource.org/license/mit
 
+#include "ecodive.h"
 
-#include <R.h>
-#include <Rinternals.h>
-#include <inttypes.h> // uint32_t, uint64_t
-#include <math.h>     // fabs, floor
-#include <string.h>   // memcpy, memset
-#include <stdlib.h>   // qsort
-#include "get.h"
-#include "memory.h"
-#include "pcg_basic.h"
-
-// Detect if pthread is available.
-#if defined __has_include
-#  if __has_include (<pthread.h>)
-#    include <pthread.h>
-#    define HAVE_PTHREAD
-#  endif
-#endif
 
 typedef void *(*pthread_func_t)(void *);
 
@@ -276,11 +260,10 @@ static pthread_func_t setup_matrix(void) {
 
 static pthread_func_t setup_dgeMatrix(void) {
   
-  SEXP sexp_dim_vec = R_do_slot(sexp_res_mtx, install("Dim"));
-  SEXP sexp_val_vec = R_do_slot(sexp_res_mtx, install("x"));
+  SEXP sexp_dim_vec = PROTECT(R_do_slot(sexp_res_mtx, install("Dim")));
+  SEXP sexp_val_vec = PROTECT(R_do_slot(sexp_res_mtx, install("x")));
   SEXP sexp_res_vec = PROTECT(duplicate(sexp_val_vec));
   R_do_slot_assign(sexp_res_mtx, install("x"), sexp_res_vec);
-  UNPROTECT(1);
   
   val_vec = REAL(sexp_val_vec);
   res_vec = REAL(sexp_res_vec);
@@ -295,7 +278,10 @@ static pthread_func_t setup_dgeMatrix(void) {
     n_otus  = INTEGER(sexp_dim_vec)[0];
   }
   
-  return setup_dense();
+  pthread_func_t rarefy_func = setup_dense();
+
+  UNPROTECT(3);
+  return rarefy_func;
 }
 
 
@@ -307,25 +293,32 @@ static pthread_func_t setup_dgeMatrix(void) {
 
 static pthread_func_t setup_slam(void) {
   
-  SEXP sexp_val_vec = get(sexp_res_mtx, "v");
+  SEXP sexp_val_vec = PROTECT(get(sexp_res_mtx, "v"));
+  SEXP sexp_i       = PROTECT(get(sexp_res_mtx, "i"));
+  SEXP sexp_j       = PROTECT(get(sexp_res_mtx, "j"));
+  SEXP sexp_nrow    = PROTECT(get(sexp_res_mtx, "nrow"));
+  SEXP sexp_ncol    = PROTECT(get(sexp_res_mtx, "ncol"));
+
   SEXP sexp_res_vec = PROTECT(duplicate(sexp_val_vec));
   set(sexp_res_mtx, "v", sexp_res_vec);
-  UNPROTECT(1);
   
   val_vec = REAL(sexp_val_vec);
   res_vec = REAL(sexp_res_vec);
   n_vals  = LENGTH(sexp_val_vec);
   
   if (margin == 1) {
-    sam_vec = INTEGER(get(sexp_res_mtx, "i"));
-    n_sams  = INTEGER(get(sexp_res_mtx, "nrow"))[0] + 1;
+    sam_vec = INTEGER(sexp_i);
+    n_sams  = INTEGER(sexp_nrow)[0] + 1;
   }
   else {
-    sam_vec = INTEGER(get(sexp_res_mtx, "j"));
-    n_sams  = INTEGER(get(sexp_res_mtx, "ncol"))[0] + 1;
[…]
