### 2026-01-16 #9082 @goldingn
Thanks for the offers to review, @equitable-equations, @CosteaPaul, and @HajkD! Thanks also @cstawitz for waiting while we lined up a second reviewer.

As @CosteaPaul is available now and is comfortable reviewing both the C and R, I'll add them as the second reviewer and kick off the review thread. You'll both see instructions in that thread on how to create your review checklist and get underway.
