### 2026-01-16 #9777 @dansmith01
@goldingn would you mind setting the version for review to "2.2.2"?
`@editorialbot set v2.2.2 as version`

Thank you @cstawitz and @CosteaPaul for agreeing to review! I look forward to your feedback.
