+++ pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   ecodive: ecodive.html
   performance: performance.html
   articles/unifrac: unifrac.html
-last_built: 2026-01-15T21:41Z
+last_built: 2026-01-16T16:47Z
 urls:
   reference: https://cmmr.github.io/ecodive/reference
   article: https://cmmr.github.io/ecodive/articles
