### 2026-01-16 d0ab0bb
Coerce counts to double before rarefaction.
