### 2026-01-20 f059ecb
`paper.md` now correctly references `vignette('performance')`.
