### 2026-01-20 3279c35
`paper.md` now references `benchmark.Rmd` article and online benchmark docs.
