### 2026-01-27 #9777 @cstawitz
@dansmith01 I am nearly done with my review so you will see some questions and issues coming in :)

I had a question about authorship as you have been the only GitHub contributor to this project, so I don't see any contributions of others to the code or the paper. Can you confirm that your two co-authors satisfy the JOSS requirements below?
_Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are._
