### 2026-01-27 cmmr/ecodive#8 @cstawitz
Suggest adding more helpful function reference
The provided documentation for diversity metrics is not very user-friendly. For example, one document "bdiv_functions.Rd" serves as function reference for all of the beta diversity metrics. The description of this entry just says "beta diversity metrics." A definition of what beta diversity is is provided in the associated vignette,  which is not even linked or mentioned in the function reference. As a user, if I seek help on a metric by typing `?aitchison` into the R console, it would be much more helpful to see documentation for just that metric, rather than having to scroll all the way down to the associated mathematical function. Further, the diversity metrics take different arguments, and the specific argument for aitchison (`pseudocount`) isn't included in the equation. The documentation for `list_metrics `is missing descriptions for some of the arguments.
