### 2026-01-27 cmmr/ecodive#9 @cstawitz
Please cite more references for the "beta diversity metric" vignette
The decision tree in the vignette "selecting a beta diversity metric" is very helpful, but lacks citations. I would expect such an authoritative resource to cite where the recommendations for different diversity metrics comes from, as I expect it draws upon a lof expert opinion. I'm not an expert in microbial ecology, so I can't verify the accuracy of the guide, but at minimum it should provide a reference list so researchers can cite why they chose a specific diversity metric.

There are also a few missing explanatory statements from this vignette. For example, the vignette recommends different alpha levels for the UniFrac method without explaining what alpha is.
