### 2026-01-27 #9777 @cstawitz
Added https://github.com/cmmr/ecodive/issues/9 and https://github.com/cmmr/ecodive/issues/8
