### 2026-01-28 #9777 @dansmith01
@cstawitz, thanks for the review and these important questions. I am happy to clarify the authorship contributions and the history of the package’s development.

### Regarding Authorship

I confirm that my co-authors satisfy JOSS authorship requirements. While I am the primary code contributor, they provided active project direction and significantly contributed to the writing, editing, and critical revision of the manuscript.

### Regarding Collaborative Development and Community Usage

The "single contributor" history reflects the fact that `ecodive` is a strategic refactoring of the diversity engine from the `rbiom` package. This code has been publicly available and in active use for several years within `rbiom`. The decision to spin it out into a standalone package was driven by community need for a lightweight, dependency-free library for these specific metrics.

While the git history is recent, the codebase has been extensively battle-tested by the community.

**Evidence of Community Feedback & Adoption:**

* **Community-Driven Optimization:** The community has actively recommended this implementation over standard tools for its performance. For example, in **[phyloseq issue 956](https://github.com/joey711/phyloseq/issues/956)**, users discussed the limitations of existing tools (`picante`) and pointed to this implementation as the preferred, faster alternative.
* **Adoption by Other Tools:** The [**`mia` package**](https://bioconductor.org/packages/mia/) (a widely used microbiome analysis suite) is currently transitioning its dependency from `rbiom` to `ecodive` to leverage these optimized functions directly (visible in their [development branch](https://github.com/microbiome/mia)).
* **Scientific Validation:** As part of `rbiom`, these functions have been successfully used in numerous independent peer-reviewed publications, including [*Obesity* (2023)](https://doi.org/10.1002/oby.23717), [*mSphere* (2024)](https://doi.org/10.1128/msphere.00825-24), and [*Frontiers in Microbiology* (2026)](https://doi.org/10.3389/fmicb.2025.1673354).

**ATIMA Web App:**
This code also serves as the backend for the [Agile Toolkit for Incisive Microbial Analyses (ATIMA)](https://atima.research.bcm.edu/). While ATIMA is currently unreleased open source, it has been cited in approximately 100 peer-reviewed publications. This JOSS paper is a critical step in formally releasing this battle-tested ecosystem to the public.

### Note on JOSS Guidelines

Finally, I would like to note that this manuscript was originally submitted in September 2025, prior to the significant JOSS guidelines update in January 2026. Consequently, the current draft lacks several now-mandatory sections, such as the "Software Design" and "Research Impact" statements. I am fully committed to updating the manuscript to meet these new standards during this review, which will further clarify the design choices and community adoption discussed above.
