### 2026-01-28 74cb135
Added citations to "Selecting a Beta Diversity Metric" vignette (#9)
