+++ vignettes/bdiv_guide.Rmd
@@ -9,35 +9,37 @@ vignette: >
 
 ## A Practical Guide to Metric Selection:<br>Aligning Method with Research Question
 
-The theoretical and mathematical properties of the various beta diversity metrics provide the foundation for making an informed choice. However, the ultimate decision must be driven by the specific research question and the nature of the dataset. This vignette provides a practical framework, including a decision tree and illustrative case studies, to guide researchers from their scientific question to an appropriate metric.
+With dozens of ways to measure the difference between microbial communities (beta diversity), choosing the "right" one can feel overwhelming. Should you use UniFrac? Bray-Curtis? Aitchison?
 
+The good news is that you don't need to master the complex mathematics behind every metric to make a robust choice. (Though if you *do* want a deep dive into the theory, **[Gloor et al. (2017)](https://doi.org/10.3389/fmicb.2017.02224)** and **[Kers and Saccenti (2022)](https://doi.org/10.3389/fmicb.2021.796025)** offer excellent open-access reviews on data compositionality and statistical power).
 
-
+For most researchers, the process relies on just a few key considerations. Instead of searching for a universally "perfect" metric, your goal is to select the tool that best aligns with your specific data properties and your biological hypothesis. This vignette cuts through the noise to provide a practical decision tree and real-world case studies, helping you match your statistical tools to your research questions.
 
 ### A Decision Tree for Metric Selection
 
-This decision tree presents a series of questions to help a researcher systematically narrow the field of 37 metrics to a small, relevant subset.
+This decision tree presents a series of questions to help a researcher systematically narrow the field of metrics to a small, relevant subset.
+
 
 
 #### 1. Do you have a reliable phylogenetic tree relating your taxa (e.g., ASVs/OTUs)?
 
 * **YES:** Your analysis can and likely should leverage phylogenetic information.
 
-    * Are you primarily interested in the presence/absence of entire evolutionary lineages (e.g., detecting the invasion of a novel phylum)? -> Start with <u>Unweighted UniFrac</u>.
+    * Are you primarily interested in the presence/absence of entire evolutionary lineages (e.g., detecting the invasion of a novel phylum)? -> Start with <u>Unweighted UniFrac</u>. As described by **[Lozupone & Knight (2005)](https://doi.org/10.1128/aem.71.12.8228-8235.2005)**, this metric is qualitative (binary) and measures the fraction of branch lengths unique to one community, making it highly sensitive to community membership.
     
-    * Are you primarily interested in shifts in the abundance of major, well-established lineages (e.g., the Firmicutes/Bacteroidetes ratio)? -> Start with <u>Weighted UniFrac</u>.
+    * Are you primarily interested in shifts in the abundance of major, well-established lineages (e.g., the Firmicutes/Bacteroidetes ratio)? -> Start with <u>Weighted UniFrac</u>. This metric weights branch lengths by abundance, thereby emphasizing dominant taxa and minimizing the noise from rare species.
     
-    * Are you interested in a robust analysis that captures changes across rare, moderate, and abundant lineages, or are you unsure? -> Use <u>Generalized UniFrac</u> (with &alpha;=0.5) as a powerful and balanced primary choice. Consider using <u>Variance-Adjusted Weighted UniFrac</u> for increased statistical power if you have uneven sample sizes.
+    * Are you interested in a robust analysis that captures changes across rare, moderate, and abundant lineages? -> Use <u>Generalized UniFrac</u>. Simulation studies by **[Chen & Li (2012)](https://doi.org/10.1093/bioinformatics/bts342)** demonstrate that Generalized UniFrac often provides more power to detect changes in moderately abundant taxa than either Unweighted or Weighted UniFrac alone.
 
[…]
