+++ authors.md
@@ -11,16 +11,16 @@
 ## Citation
 
 Source:
-[`DESCRIPTION`](https://github.com/cmmr/ecodive/blob/v2.2.2/DESCRIPTION)
+[`DESCRIPTION`](https://github.com/cmmr/ecodive/blob/main/DESCRIPTION)
 
 Smith D (2026). *ecodive: Parallel and Memory-Efficient Ecological
-Diversity Metrics*. R package version 2.2.2,
+Diversity Metrics*. R package version 2.2.3,
 <https://cmmr.github.io/ecodive/>.
 
     @Manual{,
       title = {ecodive: Parallel and Memory-Efficient Ecological Diversity Metrics},
       author = {Daniel P. Smith},
       year = {2026},
-      note = {R package version 2.2.2},
+      note = {R package version 2.2.3},
       url = {https://cmmr.github.io/ecodive/},
     }
