### 2026-01-28 98c075e
Test with integer class data
