### 2026-01-28 5282abd
Individual manual pages for each metric (#8)
