+++ R/alpha_div.r
@@ -98,99 +98,44 @@ alpha_div <- function (
 }
 
 
-
-
-
-#' Alpha Diversity Metrics
+#' Abundance-based Coverage Estimator (ACE)
 #' 
+#' A non-parametric estimator of species richness that separates features into abundant and rare groups.
 #' 
 #' @inherit documentation
-#' @name adiv_functions
-#' @family adiv_functions
+#' @family Richness metrics
+#' @seealso alpha_div
 #' 
-#' @return A numeric vector.
+#' @details
+#' The ACE metric separates features into "abundant" and "rare" groups based on a cutoff (usually 10 counts). It assumes that the presence of abundant species is certain, while the true number of rare species must be estimated.
 #' 
+#' **Equations:**
 #' 
-#' @section Formulas:
+#' \deqn{C_{ace} = 1 - \frac{F_1}{X_{rare}}}
 #' 
-#' Prerequisite: all counts are whole numbers.
+#' \deqn{\gamma_{ace}^2 = \max\left[\frac{F_{rare} \sum_{i=1}^{r}i(i-1)F_i}{C_{ace}X_{rare}(X_{rare} - 1)} - 1, 0\right]}
 #' 
-#' Given:
+#' \deqn{D_{ace} = F_{abund} + \frac{F_{rare}}{C_{ace}} + \frac{F_1}{C_{ace}}\gamma_{ace}^2}
 #' 
-#' * \eqn{n} : The number of features (e.g. species, OTUs, ASVs, etc).
-#' * \eqn{X_i} : Integer count of the \eqn{i}-th feature.
-#' * \eqn{X_T} : Total of all counts (i.e. sequencing depth). \eqn{X_T = \sum_{i=1}^{n} X_i}
-#' * \eqn{P_i} : Proportional abundance of the \eqn{i}-th feature. \eqn{P_i = X_i / X_T}
-#' * \eqn{F_1} : Number of features where \eqn{X_i = 1} (i.e. singletons).
-#' * \eqn{F_2} : Number of features where \eqn{X_i = 2} (i.e. doubletons).
-#' 
-#' |              |                                    |
-#' | :----------- | :--------------------------------- |
-#' | **Abundance-based Coverage Estimator (ACE)** <br> `ace()`         | See below. |
-#' | **Berger-Parker Index**                      <br> `berger()`      | \eqn{\max(P_i)} |
-#' | **Brillouin Index**                          <br> `brillouin()`   | \eqn{\displaystyle \frac{\ln{[(\sum_{i = 1}^{n} X_i)!]} - \sum_{i = 1}^{n} \ln{(X_i!)}}{\sum_{i = 1}^{n} X_i}} |
-#' | **Chao1**                                    <br> `chao1()`       | \eqn{\displaystyle n + \frac{(F_1)^2}{2 F_2}} |
-#' | **Faith's Phylogenetic Diversity**           <br> `faith()`       | See below. |
-#' | **Fisher's Alpha (\eqn{\alpha})**            <br> `fisher()`      | \eqn{\displaystyle \frac{n}{\alpha} = \ln{\left(1 + \frac{X_T}{\alpha}\right)}} <br> The value of \eqn{\alpha} must be solved for iteratively. |
-#' | **Gini-Simpson Index**                       <br> `simpson()`     | \eqn{1 - \sum_{i = 1}^{n} P_i^2} |
-#' | **Inverse Simpson Index**                    <br> `inv_simpson()` | \eqn{1 / \sum_{i = 1}^{n} P_i^2} |
[…]
