+++ man/fisher.Rd
@@ -1,0 +1,61 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/alpha_div.r
+\name{fisher}
+\alias{fisher}
+\title{Fisher's Alpha}
+\usage{
+fisher(counts, digits = 3L, margin = 1L, cpus = n_cpus())
+}
+\arguments{
+\item{counts}{A numeric matrix of count data where each column is a
+feature, and each row is a sample. Any object coercible with
+\code{as.matrix()} can be given here, as well as \code{phyloseq}, \code{rbiom},
+\code{SummarizedExperiment}, and \code{TreeSummarizedExperiment} objects. For
+optimal performance with very large datasets, see the guide in
+\code{vignette('performance')}.}
+
+\item{digits}{Precision of the returned values, in number of decimal
+places. E.g. the default \code{digits=3} could return \code{6.392}.}
+
+\item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
+your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
+\code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, or
+\code{TreeSummarizedExperiment} object. Default: \code{1L}}
+
+\item{cpus}{How many parallel processing threads should be used. The
+default, \code{n_cpus()}, will use all logical CPU cores.}
+}
+\description{
+A parametric diversity index assuming species abundances follow a log-series distribution.
+}
+\details{
+Fisher's Alpha (\eqn{\alpha}) is the parameter in the equation:
+\deqn{\frac{n}{\alpha} = \ln{\left(1 + \frac{X_T}{\alpha}\right)}}
+
+Where:
+\itemize{
+\item \eqn{n} : The number of features.
+\item \eqn{X_T} : Total of all counts (sequencing depth).
+}
+
+The value of \eqn{\alpha} is solved for iteratively.
+
+\strong{Parameter: digits}
+The precision (number of decimal places) to use when solving the equation.
+}
+\examples{
+    fisher(ex_counts)
+}
+\references{
+Fisher, R. A., Corbet, A. S., & Williams, C. B. (1943). The relation between the number of species and the number of individuals in a random sample of an animal population. \emph{Journal of Animal Ecology}, 12, 42-58. \doi{10.2307/1411}
+}
+\seealso{
+alpha_div
+
+Other Diversity metrics: 
+\code{\link{brillouin}()},
+\code{\link{inv_simpson}()},
+\code{\link{shannon}()},
+\code{\link{simpson}()}
+}
+\concept{Diversity metrics}

[…]
