+++ reference/chord.md
@@ -1,0 +1,105 @@
+# Chord distance
+
+Euclidean distance between normalized vectors.
+
+## Usage
+
+``` r
+chord(counts, margin = 1L, pairs = NULL, cpus = n_cpus())
+```
+
+## Arguments
+
+- counts:
+
+  A numeric matrix of count data where each column is a feature, and
+  each row is a sample. Any object coercible with
+  [`as.matrix()`](https://rdrr.io/r/base/matrix.html) can be given here,
+  as well as `phyloseq`, `rbiom`, `SummarizedExperiment`, and
+  `TreeSummarizedExperiment` objects. For optimal performance with very
+  large datasets, see the guide in
+  [`vignette('performance')`](https://cmmr.github.io/ecodive/articles/performance.md).
+
+- margin:
+
+  If your samples are in the matrix's rows, set to `1L`. If your samples
+  are in columns, set to `2L`. Ignored when `counts` is a `phyloseq`,
+  `rbiom`, `SummarizedExperiment`, or `TreeSummarizedExperiment` object.
+  Default: `1L`
+
+- pairs:
+
+  Which combinations of samples should distances be calculated for? The
+  default value (`NULL`) calculates all-vs-all. Provide a numeric or
+  logical vector specifying positions in the distance matrix to
+  calculate. See examples.
+
+- cpus:
+
+  How many parallel processing threads should be used. The default,
+  [`n_cpus()`](https://cmmr.github.io/ecodive/reference/n_cpus.md), will
+  use all logical CPU cores.
+
+## Details
+
+The Chord distance is defined as: \$\$\sqrt{\sum\_{i=1}^{n}
+\left(\frac{X_i}{\sqrt{\sum\_{j=1}^{n} X_j^2}} -
+\frac{Y_i}{\sqrt{\sum\_{j=1}^{n} Y_j^2}}\right)^2}\$\$
+
+Where:
+
+- \\X_i\\, \\Y_i\\ : Absolute counts of the \\i\\-th feature.
+
+- \\n\\ : The number of features.
+
+Base R Equivalent:
+
+    sqrt(sum(((x / sqrt(sum(x ^ 2))) - (y / sqrt(sum(y ^ 2))))^2))
+
+## References
+
+Orlóci, L. (1967). An agglomerative method for classification of plant
+communities. *Journal of Ecology*, 55(1), 193-206.
+[doi:10.2307/2257725](https://doi.org/10.2307/2257725)
+
+## See also
+
+beta_div
+
+Other Abundance metrics:
+[`aitchison()`](https://cmmr.github.io/ecodive/reference/aitchison.md),
+[`bhattacharyya()`](https://cmmr.github.io/ecodive/reference/bhattacharyya.md),
+[`bray()`](https://cmmr.github.io/ecodive/reference/bray.md),
+[`canberra()`](https://cmmr.github.io/ecodive/reference/canberra.md),
+[`chebyshev()`](https://cmmr.github.io/ecodive/reference/chebyshev.md),
+[`clark()`](https://cmmr.github.io/ecodive/reference/clark.md),
+[`divergence()`](https://cmmr.github.io/ecodive/reference/divergence.md),
+[`euclidean()`](https://cmmr.github.io/ecodive/reference/euclidean.md),
+[`gower()`](https://cmmr.github.io/ecodive/reference/gower.md),
+[`hellinger()`](https://cmmr.github.io/ecodive/reference/hellinger.md),
+[`horn()`](https://cmmr.github.io/ecodive/reference/horn.md),
+[`jensen()`](https://cmmr.github.io/ecodive/reference/jensen.md),
[…]
