+++ R/alpha_div.r
@@ -104,7 +104,7 @@ alpha_div <- function (
 #' 
 #' @inherit documentation
 #' @family Richness metrics
-#' @seealso alpha_div
+#' @seealso `alpha_div()`, `vignette('adiv')`
 #' 
 #' @details
 #' The ACE metric separates features into "abundant" and "rare" groups based on a cutoff (usually 10 counts). It assumes that the presence of abundant species is certain, while the true number of rare species must be estimated.
@@ -152,7 +152,7 @@ ace <- function (counts, cutoff = 10L, margin = 1L, cpus = n_cpus()) {
 #' 
 #' @inherit documentation
 #' @family Dominance metrics
-#' @seealso alpha_div
+#' @seealso `alpha_div()`, `vignette('adiv')`
 #' 
 #' @details
 #' The Berger-Parker index is defined as the proportional abundance of the most dominant feature:
@@ -185,7 +185,7 @@ berger <- function (counts, norm = 'percent', margin = 1L, cpus = n_cpus()) {
 #' 
 #' @inherit documentation
 #' @family Diversity metrics
-#' @seealso alpha_div
+#' @seealso `alpha_div()`, `vignette('adiv')`
 #' 
 #' @details
 #' The Brillouin index is defined as:
@@ -223,7 +223,7 @@ brillouin <- function (counts, margin = 1L, cpus = n_cpus()) {
 #' 
 #' @inherit documentation
 #' @family Richness metrics
-#' @seealso alpha_div
+#' @seealso `alpha_div()`, `vignette('adiv')`
 #' 
 #' @details
 #' The Chao1 estimator uses the ratio of singletons to doubletons to estimate the number of missing species:
@@ -261,7 +261,7 @@ chao1 <- function (counts, margin = 1L, cpus = n_cpus()) {
 #' 
 #' @inherit documentation
 #' @family Phylogenetic metrics
-#' @seealso alpha_div
+#' @seealso `alpha_div()`, `vignette('adiv')`
 #' 
 #' @details
 #' Faith's PD is defined as:
@@ -293,7 +293,7 @@ faith <- function (counts, tree = NULL, margin = 1L, cpus = n_cpus()) {
 #' 
 #' @inherit documentation
 #' @family Diversity metrics
-#' @seealso alpha_div
+#' @seealso `alpha_div()`, `vignette('adiv')`
 #' 
 #' @details
 #' Fisher's Alpha (\eqn{\alpha}) is the parameter in the equation:
@@ -330,7 +330,7 @@ fisher <- function (counts, digits = 3L, margin = 1L, cpus = n_cpus()) {
 #' 
 #' @inherit documentation
 #' @family Diversity metrics
-#' @seealso alpha_div
+#' @seealso `alpha_div()`, `vignette('adiv')`
 #' 
 #' @details
 #' The Inverse Simpson index is defined as:
@@ -364,7 +364,7 @@ inv_simpson <- function (counts, norm = 'percent', margin = 1L, cpus = n_cpus())
 #' 
 #' @inherit documentation
 #' @family Richness metrics
-#' @seealso alpha_div
+#' @seealso `alpha_div()`, `vignette('adiv')`
 #' 
 #' @details
 #' Margalef's index is defined as:
@@ -381,6 +381,8 @@ inv_simpson <- function (counts, norm = 'percent', margin = 1L, cpus = n_cpus())
 #' 
 #' @references
 #' Margalef, R. (1958). Information theory in ecology. *General Systems*, 3, 36-71.
+#' 
[…]
