+++ R/beta_div.r
@@ -151,7 +151,7 @@ beta_div <- function (
 #' 
 #' @inherit documentation
 #' @family Abundance metrics
-#' @seealso beta_div
+#' @seealso `beta_div()`, `vignette('bdiv')`, `vignette('bdiv_guide')`
 #' 
 #' @details
 #' The Aitchison distance is defined as:
@@ -176,6 +176,8 @@ beta_div <- function (
 #' @references
 #' Aitchison, J. (1986). The statistical analysis of compositional data. Chapman and Hall. \doi{10.1007/978-94-009-4109-3}
 #' 
+#' Aitchison, J. (1982). The statistical analysis of compositional data. *Journal of the Royal Statistical Society: Series B (Methodological)*, 44(2), 139-160. \doi{10.1111/j.2517-6161.1982.tb01195.x}
+#'
 #' @export
 #' @examples
 #'     aitchison(ex_counts, pseudocount = 1)
@@ -194,7 +196,7 @@ aitchison <- function (counts, pseudocount = NULL, margin = 1L, pairs = NULL, cp
 #' 
 #' @inherit documentation
 #' @family Abundance metrics
-#' @seealso beta_div
+#' @seealso `beta_div()`, `vignette('bdiv')`, `vignette('bdiv_guide')`
 #' 
 #' @details
 #' The Bhattacharyya distance is defined as:
@@ -228,7 +230,7 @@ bhattacharyya <- function (counts, norm = 'percent', margin = 1L, pairs = NULL, 
 #' 
 #' @inherit documentation
 #' @family Abundance metrics
-#' @seealso beta_div
+#' @seealso `beta_div()`, `vignette('bdiv')`, `vignette('bdiv_guide')`
 #' 
 #' @details
 #' The Bray-Curtis dissimilarity is defined as:
@@ -262,7 +264,7 @@ bray <- function (counts, norm = 'percent', margin = 1L, pairs = NULL, cpus = n_
 #' 
 #' @inherit documentation
 #' @family Abundance metrics
-#' @seealso beta_div
+#' @seealso `beta_div()`, `vignette('bdiv')`, `vignette('bdiv_guide')`
 #' 
 #' @details
 #' The Canberra distance is defined as:
@@ -296,7 +298,7 @@ canberra <- function (counts, norm = 'percent', margin = 1L, pairs = NULL, cpus 
 #' 
 #' @inherit documentation
 #' @family Abundance metrics
-#' @seealso beta_div
+#' @seealso `beta_div()`, `vignette('bdiv')`, `vignette('bdiv_guide')`
 #' 
 #' @details
 #' The Chebyshev distance is defined as:
@@ -329,7 +331,7 @@ chebyshev <- function (counts, norm = 'percent', margin = 1L, pairs = NULL, cpus
 #' 
 #' @inherit documentation
 #' @family Abundance metrics
-#' @seealso beta_div
+#' @seealso `beta_div()`, `vignette('bdiv')`, `vignette('bdiv_guide')`
 #' 
 #' @details
 #' The Chord distance is defined as:
@@ -365,7 +367,7 @@ chord <- function (counts, margin = 1L, pairs = NULL, cpus = n_cpus()) {
 #' 
 #' @inherit documentation
 #' @family Abundance metrics
-#' @seealso beta_div
+#' @seealso `beta_div()`, `vignette('bdiv')`, `vignette('bdiv_guide')`
 #' 
 #' @details
 #' Clark's divergence distance is defined as:
@@ -399,7 +401,7 @@ clark <- function (counts, norm = 'percent', margin = 1L, pairs = NULL, cpus = n
 #' 
 #' @inherit documentation
[…]

[…]
