### 2026-01-29 b50a0a8
list rendering for `@param norm`.
