+++ R/documentation.r
@@ -38,12 +38,13 @@
 #'        the smallest non-zero value in `counts`.
 #' 
 #' @param norm   Normalize the incoming counts. Options are:
-#'        * `'percent'`: Relative abundance (sample abundances sum to 1).
-#'        * `'binary'`: Unweighted presence/absence (each count is either 0 or 1).
-#'        * `'clr'`: Centered log ratio.
-#'        * `'none'`: No transformation.
-#'        
-#'        Default: `'percent'`, which is the expected input for these formulas.
+#'   
+#'   * `'percent'`: Relative abundance (sample abundances sum to 1).
+#'   * `'binary'`: Unweighted presence/absence (each count is either 0 or 1).
+#'   * `'clr'`: Centered log ratio.
+#'   * `'none'`: No transformation.
+#'   
+#'   Default: `'percent'`, which is the expected input for these formulas.
 #' 
 #' @param margin   If your samples are in the matrix's rows, set to `1L`. If 
 #'        your samples are in columns, set to `2L`. Ignored when `counts` is a 
