+++ man/documentation.Rd
@@ -35,13 +35,14 @@ taking \code{log(0)} for unobserved features. The default, \code{NULL}, selects
 the smallest non-zero value in \code{counts}.}
 
 \item{norm}{Normalize the incoming counts. Options are:
-* \code{'percent'}: Relative abundance (sample abundances sum to 1).
-* \code{'binary'}: Unweighted presence/absence (each count is either 0 or 1).
-* \code{'clr'}: Centered log ratio.
-* \code{'none'}: No transformation.
+\itemize{
+\item \code{'percent'}: Relative abundance (sample abundances sum to 1).
+\item \code{'binary'}: Unweighted presence/absence (each count is either 0 or 1).
+\item \code{'clr'}: Centered log ratio.
+\item \code{'none'}: No transformation.
+}
 
-\if{html}{\out{<div class="sourceCode">}}\preformatted{   Default: `'percent'`, which is the expected input for these formulas.
-}\if{html}{\out{</div>}}}
+Default: \code{'percent'}, which is the expected input for these formulas.}
 
 \item{margin}{If your samples are in the matrix's rows, set to \code{1L}. If
 your samples are in columns, set to \code{2L}. Ignored when \code{counts} is a
