### 2026-01-29 #9777 @cstawitz
Thank you for the additional clarification, and for your responses to the issues regarding documentation on the package repository.

I agree it would be very helpful to update the manuscript to add the additional citations and contribution history that you denote above. The improved documentation, additional citations, and community history described above were my main points of feedback. As someone outside of the microbial ecology community, I wasn't aware of the `rbiom` history and citations, so it would be very helpful to describe those more clearly in the manuscript. I'd be happy to take another look once the manuscript is updated, but expect the edits you propose to address most of my feedback.
