### 2026-01-29 bd0b6dd
Manuscript revised per #10 and #11 and updated JOSS guidelines.
