+++ paper/paper.bib
@@ -1,54 +1,3 @@
-
-
-@manual{abdiv,
-  title  = {abdiv: Alpha and Beta Diversity Measures},
-  author = {Bittinger, Kyle},
-  year   = {2020},
-  note   = {R package version 0.2.0},
-  url    = {https://CRAN.R-project.org/package=abdiv},
-  doi    = {10.32614/CRAN.package.abdiv}
-}
-
-@article{adiv,
-  title   = {adiv: An r package to analyse biodiversity in ecology},
-  author  = {Sandrine Pavoine},
-  year    = {2020},
-  journal = {Methods in Ecology and Evolution},
-  volume  = {11},
-  pages   = {1106-1112},
-  doi     = {10.1111/2041-210X.13430}
-}
-
-@article{ampvis2,
-  title   = {ampvis2: an R package to analyse and visualise 16S rRNA amplicon data},
-  author  = {Andersen, Kasper S. and Kirkegaard, Rasmus H. and Karst, Søren M. and Albertsen, Mads},
-  year    = {2018},
-  journal = {bioRxiv},
-  doi     = {10.1101/299537}
-}
-
-@article{ecodist,
-  title = {The ecodist package for dissimilarity-based analysis of ecological data},
-  author = {Sarah C. Goslee and Dean L. Urban},
-  year = {2007},
-  journal = {Journal of Statistical Software},
-  volume = {22},
-  issue = {7},
-  pages = {1-19},
-  doi = {10.18637/jss.v022.i07}
-}
-
-@article{entropart,
-  title   = {entropart: An R Package to Measure and Partition Diversity},
-  author  = {Eric Marcon and Bruno Herault},
-  journal = {Journal of Statistical Software},
-  year    = {2015},
-  volume  = {67},
-  number  = {8},
-  pages   = {1-26},
-  doi     = {10.18637/jss.v067.i08}
-}
-
 @manual{GUniFrac,
   title  = {GUniFrac: Generalized UniFrac Distances, Distance-Based
     Multivariate Methods and Feature-Based Univariate Methods for
@@ -58,46 +7,20 @@
   doi    = {10.32614/CRAN.package.GUniFrac}
 }
 
-@manual{labdsv,
-  title = {labdsv: Ordination and Multivariate Analysis for Ecology},
-  author = {David W. Roberts},
-  year = {2025},
-  doi = {10.32614/CRAN.package.labdsv}
-}
-
-
-@manual{parallelDist,
-  title = {parallelDist: Parallel Distance Matrix Computation using Multiple Threads},
-  author = {Alexander Eckert},
-  year = {2022},
-  doi = {10.32614/CRAN.package.parallelDist}
-}
-
-
 @article{philentropy,
-  title = {Philentropy: Information Theory and Distance Quantification with R},
-  author = {Drost HG},
+  title   = {Philentropy: Information Theory and Distance Quantification with R},
+  author  = {Hajk-Georg Drost},
   journal = {Journal of Open Source Software},
-  year = {2018},
-  volume = {3},
-  number = {26},
-  pages = {765},
-  doi = {10.21105/joss.00765},
-}
-
-
-@article{phyloregion,
-  author  = {Barnabas H. Daru and Piyal Karunarathne and Klaus Schliep},
-  title   = {phyloregion: R package for biogeographic regionalization and macroecology},
-  year    = {2020},
-  journal = {Methods in Ecology and Evolution},
-  volume  = {11},
-  pages   = {1483-1491},
-  doi     = {10.1111/2041-210X.13478}
+  year    = {2018},
+  volume  = {3},
+  number  = {26},
+  pages   = {765},
+  doi     = {10.21105/joss.00765},
 }
 
 @article{phyloseq,
[…]
