+++ paper/paper.md
@@ -11,7 +11,7 @@ authors:
 - name: Joseph F Petrosino
   orcid: "0000-0002-4046-6898"
   affiliation: "1, 2"
-date: "19 September 2025"
+date: "29 January 2026"
 output:
   word_document: default
   pdf_document: default
@@ -31,110 +31,51 @@ affiliations:
 bibliography: paper.bib
 ---
 
-
-
 # Summary
 
-Characterizing the composition of biological communities is a fundamental task
-in ecology, but the calculations involved can be computationally prohibitive
-when applied to large datasets. `ecodive` is an R package that addresses this
-challenge by providing highly optimized implementations of 50 ecological
-metrics for alpha-diversity (within-sample richness and evenness) and
-beta-diversity (between-sample dissimilarity). These metrics can incorporate
-species counts, relative abundances, and evolutionary relationships, providing a
-multi-faceted view of ecological structure. By leveraging a compiled C library
-with pthreads for parallelization, `ecodive` delivers substantial performance
-gains in both speed and memory usage, enabling researchers to analyze large
-datasets quickly and efficiently.
-
-
+Understanding the complexity of biological communities—whether bacteria in the human gut, trees in a forest, or plankton in the ocean—is a central goal of ecology. Researchers quantify this complexity using "diversity metrics," which describe the variety of species within a single site (alpha diversity) or the differences in composition between two sites (beta diversity). `ecodive` is an R package designed to calculate these metrics efficiently. It bridges the gap between complex ecological theory and practical data analysis, providing researchers with a unified toolset to process large-scale datasets that were previously computationally prohibitive. By leveraging parallel processing and optimized memory management, `ecodive` enables rapid, high-throughput analysis of microbial and macro-ecological communities.
 
 # Statement of Need
 
-A primary challenge in large-scale ecological analysis is the computational
-complexity of beta diversity calculations. These algorithms exhibit $O(n^2)$
-complexity, meaning their computational cost scales quadratically with the
-number of samples ($n$). As microbiome and ecological studies grow to include
-thousands of samples, this quadratic scaling creates a significant bottleneck,
-demanding immense processing time and memory.
+A primary challenge in modern ecological analysis is the management of high-dimensional data. As sequencing technologies improve, datasets are growing to include thousands of samples and tens of thousands of unique taxa. Beta diversity calculations, which involve comparing every sample to every other sample, exhibit $O(n^2)$ complexity. This quadratic scaling creates a significant bottleneck; a dataset that doubles in size requires four times the processing power, often overwhelming standard desktop computers.
 
-A second challenge is the fragmentation of diversity metrics across numerous R
-packages. Researchers often need to install and manage a suite of dependencies
-to access the full range of metrics required for a comprehensive analysis,
-leading to potential version conflicts and a disjointed workflow.
+Furthermore, the software landscape for ecological metrics is fragmented. A researcher needing to calculate a specific set of indices—for example, Faith's Phylogenetic Diversity, Bray-Curtis dissimilarity, and UniFrac distances—often must install and manage multiple R packages (`picante`, `vegan`, `GUniFrac`), each with different dependencies, input formats, and performance limitations.
 
-`ecodive` addresses both of these critical needs. First, it provides a highly
-optimized, parallelized C-based engine that dramatically reduces the time and
-memory required by these algorithms, enabling the analysis of much larger
-datasets. Second, it consolidates a vast collection of alpha and beta diversity
-metrics into a single, dependency-free package. By solving the dual problems of
[…]
