### 2026-01-29 e2255d6
manuscript tweak
