### 2026-01-29 bf6b2dd
more minor manuscript tweaks
