### 2026-01-29 #9777 @dansmith01
@cstawitz, thank you for the follow-up. I have updated the manuscript to include the additional citations and the development history regarding `rbiom` as discussed. The updated vignettes and function documentation are also now live in the development version and on the [package website](https://cmmr.github.io/ecodive/).

I have replied to the individual issues you opened with specific details on those fixes. I have left those issues open so you can review the changes and close them if you feel they have been adequately addressed.
