### 2026-01-29 7f6d79c
Minkowski's special cases
