+++ R/alpha_div.r
@@ -306,6 +306,7 @@ faith <- function (counts, tree = NULL, margin = 1L, cpus = n_cpus()) {
 #' The value of \eqn{\alpha} is solved for iteratively.
 #' 
 #' **Parameter: digits**
+#' 
 #' The precision (number of decimal places) to use when solving the equation.
 #' 
 #' @references
