### 2026-01-31 #9777 @CosteaPaul
Code looks good, both the C and the R bits. The speedup is also very nice 👍

I have added some issues to the repo, mostly concerning defaults of some of the functions and clarity of use. As a general principle, i want the functions i use to do the most obvious thing that the name suggests they will do. When they start messing with the data passed in as default behavior, this can lead to a lot of usage errors.

As far the review is concerned, these issues should be seen as somewhere between "would be nice to have" and "pedantic".
