### 2026-02-02 9f15c15
Refactor rarefy(): scope reduction and explicit drop handling (#12)
