+++ tests/testthat/test-rarefy.r
@@ -1,35 +1,39 @@
 test_that("rarefaction", {
   
   # Basic return types
-  
   expect_true(is.matrix(rarefy(counts)))
   expect_true(is.list(rarefy(counts, times = 2)))
   
+  # Check reproducibility (seed) and default behavior
   expect_identical(
-    object   = as.vector(rarefy(counts, seed = 1, drop = FALSE)), 
+    object   = as.vector(rarefy(counts, seed = 1, drop = FALSE, warn = FALSE)), 
     expected = c(0,0,0,0,0,4,3,10,5,4,5,3,2,0,0,0,6,5,5,0) )
   
   expect_identical(
-    object   = as.vector(rarefy(counts, seed = 2, drop = FALSE)), 
+    object   = as.vector(rarefy(counts, seed = 2, drop = FALSE, warn = FALSE)), 
     expected = c(0,0,0,0,0,7,5,9,5,2,2,4,2,0,0,0,6,4,6,0) )
   
+  # Test Auto-Depth Selection 
+  # Default auto-selection (depth=13) retains all samples in this dataset 
+  # (A=13, B=18, C=21, D=15), so NO warning is issued.
   expect_identical(
-    object   = rarefy(counts, n_samples = 0), 
-    expected = rarefy(counts, n_samples = 4) )
+    object   = rarefy(counts, depth = NULL, warn = TRUE), 
+    expected = rarefy(counts, depth = 13, warn = TRUE) 
+  )
   
-  expect_identical(
-    object   = rarefy(counts, n_samples = 0.5), 
-    expected = rarefy(counts, n_samples = 2) )
+  # Test Auto-Depth Fallback (Branch coverage for validate_depth fallback)
+  # Create a matrix where 10% yield is impossible unless we pick max depth
+  # e.g., one huge sample and many tiny ones.
+  # This SHOULD warn because the tiny samples (1, 1) are < 100.
+  tiny_counts <- matrix(c(100, 1, 1), ncol=1)
+  expect_warning(
+    res <- rarefy(tiny_counts, depth = NULL, warn = TRUE), 
+    regexp = "dropped"
+  )
+  expect_equal(sum(res), 100) # Should pick max depth (100) and drop others
   
-  expect_identical(
-    object   = rarefy(counts, n_samples = 3), 
-    expected = rarefy(counts, n_samples = -1) )
   
-  expect_identical(
-    object   = rarefy(counts, depth = 0.1), 
-    expected = rarefy(counts, depth = 13) )
-  
-  expect_silent(m <- rarefy(big_mtx, 15))
+  expect_silent(m <- rarefy(big_mtx, 15, warn = FALSE))
   expect_in(rowSums(m), range(rowSums(m)))
   
   # Test times parameter
@@ -38,27 +42,53 @@ test_that("rarefaction", {
   expect_true(is.matrix(r_list[[1]]))
   expect_false(identical(r_list[[1]], r_list[[2]]))
   
-  # Test drop parameter
+  # -----------------------------------------------------------------------
+  # Test drop parameter & Warning Logic
+  # -----------------------------------------------------------------------
   
-  # Rarefy to a high depth to create zero-sum rows/cols
-  r_drop_F <- rarefy(counts, depth = 15, drop = FALSE, seed = 1)
+  # Case 1: drop = FALSE (Retain Original Counts)
+  # Rarefy to depth 15. Sample A (13) < 15.
+  # They should be returned UNMODIFIED.
[…]
