+++ R/rarefy.r
@@ -3,38 +3,54 @@
 
 
 
-#' Rarefy OTU counts.
+#' Rarefy Observation Counts
 #' 
-#' Sub-sample OTU observations such that all samples have an equal number.
-#' If called on data with non-integer abundances, values will be re-scaled to 
-#' integers between 1 and `depth` such that they sum to `depth`.
+#' Sub-sample observations from a feature table such that all samples have the 
+#' same library size (depth). This is performed via random sampling without 
+#' replacement.
+#' 
+#' @details
+#' **Auto-Depth Selection**\cr
+#' If `depth` is `NULL`, the function defaults to the highest depth that retains 
+#' at least 10% of the total observations in the dataset.
+#' 
+#' **Dropping vs. Retaining Samples**\cr
+#' If a sample has fewer observations than the specified `depth`:
+#' 
+#' * `drop = TRUE` (Default): The sample is removed from the output matrix.
+#' * `drop = FALSE`: The sample is returned **unmodified** (with its original 
+#'   counts). It is *not* rarefied or zeroed out.
+#' 
+#' **Zero-Sum Features**\cr
+#' Features (OTUs, ASVs, Genes) that lose all observations during rarefaction 
+#' are **always retained** as columns/rows of zeros. This ensures the output 
+#' matrix dimensions remain consistent with the input (barring dropped samples).
 #' 
 #' @inherit documentation
 #' 
-#' @param depth   How many observations to keep per sample. When 
-#'        `0 < depth < 1`, it is taken as the minimum percentage of the 
-#'        dataset's observations to keep. Ignored when `n_samples` is 
-#'        specified. Default: `0.1`
-#'
-#' @param n_samples   The number of samples to keep. When `0 < n_samples < 1`, 
-#'        it is taken as the percentage of samples to keep. If negative, that 
-#'        number of samples is dropped. If `0`, all samples are kept. 
-#'        If `NULL`, then `depth` is used instead. Default: `NULL`
-#'     
-#' @param seed   An integer seed for randomizing which observations to keep or 
-#'        drop. If you need to create different random rarefactions of the same 
-#'        data, set the seed to a different number each time. Default: `0`
-#'     
-#' @param times   How many independent rarefactions to perform. If set, 
-#'        `rarefy()` will return a list of matrices. The seeds for each matrix
-#'        will be sequential, starting from `seed`. Default: `NULL`
-#'     
-#' @param drop   Drop rows and columns with zero observations after rarefying.
-#'        Default: `TRUE`
+#' @param counts  A numeric matrix or sparse matrix object (e.g., `dgCMatrix`).
+#'        Counts must be integers (non-integer counts will be cast to integers).
 #' 
+#' @param depth   The number of observations to keep per sample. If `NULL` 
+#'        (the default), a depth is auto-selected to maximize data retention.
+#'     
+#' @param seed    An integer seed for the random number generator. Providing 
+#'        the same seed guarantees reproducible results. Default: `0`
+#'     
+#' @param times   The number of independent rarefactions to perform. If set, 
+#'        returns a list of matrices. Seeds for subsequent iterations are 
+#'        sequential (`seed`, `seed + 1`, ...). Default: `NULL`
+#'     
+#' @param drop    Logical. If `TRUE`, samples with fewer than `depth` 
[…]
