+++ R/documentation.r
@@ -46,9 +46,9 @@
 #'   
 #'   Default: `'percent'`, which is the expected input for these formulas.
 #' 
-#' @param margin   If your samples are in the matrix's rows, set to `1L`. If 
-#'        your samples are in columns, set to `2L`. Ignored when `counts` is a 
-#'        special object class (e.g. `phyloseq`). Default: `1L`
+#' @param margin  The margin containing samples. `1` if samples are rows, 
+#'        `2` if samples are columns. Ignored when `counts` is a special object 
+#'        class (e.g. `phyloseq`). Default: `1`
 #' 
 #' @param tree   A `phylo`-class object representing the phylogenetic tree for 
 #'        the OTUs in `counts`. The OTU identifiers given by `colnames(counts)` 

[…]
