### 2026-02-04 857aa24
Refactor norm parameter defaults and update diversity formulas (#13)
