+++ R/alpha_div.r
@@ -103,6 +103,8 @@ alpha_div <- function (
 #' A non-parametric estimator of species richness that separates features into abundant and rare groups.
 #' 
 #' @inherit documentation
+#' @inherit adiv_assert_integer
+#' 
 #' @family Richness metrics
 #' @seealso `alpha_div()`, `vignette('adiv')`
 #' 
@@ -151,6 +153,8 @@ ace <- function (counts, cutoff = 10L, margin = 1L, cpus = n_cpus()) {
 #' A measure of the numerical importance of the most abundant species.
 #' 
 #' @inherit documentation
+#' @inherit adiv_percent_normalized
+#' 
 #' @family Dominance metrics
 #' @seealso `alpha_div()`, `vignette('adiv')`
 #' 
@@ -163,7 +167,8 @@ ace <- function (counts, cutoff = 10L, margin = 1L, cpus = n_cpus()) {
 #' 
 #' **Base R Equivalent:**
 #' ```r
-#' max(x / sum(x))
+#' p <- x / sum(x)
+#' max(p)
 #' ```
 #' 
 #' @references
@@ -172,9 +177,11 @@ ace <- function (counts, cutoff = 10L, margin = 1L, cpus = n_cpus()) {
 #' @export
 #' @examples
 #'     berger(ex_counts)
-berger <- function (counts, norm = 'percent', margin = 1L, cpus = n_cpus()) {
+berger <- function (counts, margin = 1L, cpus = n_cpus()) {
   
+  norm <- 'percent'
   validate_args()
+  
   .Call(C_alpha_div, ADIV_BERGER, counts, margin, norm, cpus, NULL)
 }
 
@@ -184,6 +191,8 @@ berger <- function (counts, norm = 'percent', margin = 1L, cpus = n_cpus()) {
 #' A diversity index derived from information theory, appropriate for fully censused communities.
 #' 
 #' @inherit documentation
+#' @inherit adiv_assert_integer
+#' 
 #' @family Diversity metrics
 #' @seealso `alpha_div()`, `vignette('adiv')`
 #' 
@@ -222,6 +231,8 @@ brillouin <- function (counts, margin = 1L, cpus = n_cpus()) {
 #' A non-parametric estimator of the lower bound of species richness.
 #' 
 #' @inherit documentation
+#' @inherit adiv_assert_integer
+#' 
 #' @family Richness metrics
 #' @seealso `alpha_div()`, `vignette('adiv')`
 #' 
@@ -260,6 +271,8 @@ chao1 <- function (counts, margin = 1L, cpus = n_cpus()) {
 #' Calculates the sum of the branch lengths for all species present in a sample.
 #' 
 #' @inherit documentation
+#' @inherit adiv_binary_normalized
+#' 
 #' @family Phylogenetic metrics
 #' @seealso `alpha_div()`, `vignette('adiv')`
 #' 
@@ -292,6 +305,8 @@ faith <- function (counts, tree = NULL, margin = 1L, cpus = n_cpus()) {
 #' A parametric diversity index assuming species abundances follow a log-series distribution.
 #' 
 #' @inherit documentation
+#' @inherit adiv_assert_integer
+#' 
 #' @family Diversity metrics
 #' @seealso `alpha_div()`, `vignette('adiv')`
 #' 
@@ -330,6 +345,8 @@ fisher <- function (counts, digits = 3L, margin = 1L, cpus = n_cpus()) {
 #' A transformation of the Simpson index that represents the "effective number of species".
 #' 
 #' @inherit documentation
+#' @inherit adiv_percent_normalized
[…]
