+++ R/beta_div.r
@@ -38,7 +38,12 @@ V_UNIFRAC <- 5L
 #' Beta Diversity Wrapper Function
 #' 
 #' @inherit documentation
+#' 
 #' @name beta_div
+#'        
+#' @param counts   A numeric matrix of count data (samples \eqn{\times} features). 
+#'        Typically contains absolute abundances (integer counts), though 
+#'        proportions are also accepted by some diversity metrics.
 #' 
 #' @param metric   The name of a beta diversity metric. One of `c('aitchison',
 #'   'bhattacharyya', 'bray', 'canberra', 'chebyshev', 'chord', 'clark',
@@ -99,9 +104,6 @@ V_UNIFRAC <- 5L
 #' 
 #' 
 #' 
-#' 
-#' 
-#' 
 #' **Flexible name matching**
 #' 
 #' Case insensitive and partial matching. Any runs of non-alpha characters are
@@ -129,7 +131,7 @@ V_UNIFRAC <- 5L
 beta_div <- function (
     counts, 
     metric, 
-    norm        = 'percent', 
+    norm        = 'none', 
     power       = 1.5, 
     pseudocount = NULL, 
     alpha       = 0.5, 
@@ -150,6 +152,7 @@ beta_div <- function (
 #' Calculates the Euclidean distance between centered log-ratio (CLR) transformed abundances.
 #' 
 #' @inherit documentation
+#' 
 #' @family Abundance metrics
 #' @seealso `beta_div()`, `vignette('bdiv')`, `vignette('bdiv_guide')`
 #' 
@@ -196,6 +199,8 @@ aitchison <- function (counts, pseudocount = NULL, margin = 1L, pairs = NULL, cp
 #' Measures the similarity of two probability distributions.
 #' 
 #' @inherit documentation
+#' @inherit bdiv_percent_normalized
+#' 
 #' @family Abundance metrics
 #' @seealso `beta_div()`, `vignette('bdiv')`, `vignette('bdiv_guide')`
 #' 
@@ -209,7 +214,9 @@ aitchison <- function (counts, pseudocount = NULL, margin = 1L, pairs = NULL, cp
 #' 
 #' Base R Equivalent: 
 #' ```r
-#' -log(sum(sqrt(x * y)))
+#' p <- x / sum(x)
+#' q <- y / sum(y)
+#' -log(sum(sqrt(p * q)))
 #' ```
 #' 
 #' @references
@@ -218,9 +225,11 @@ aitchison <- function (counts, pseudocount = NULL, margin = 1L, pairs = NULL, cp
 #' @export
 #' @examples
 #'     bhattacharyya(ex_counts)
-bhattacharyya <- function (counts, norm = 'percent', margin = 1L, pairs = NULL, cpus = n_cpus()) {
+bhattacharyya <- function (counts, margin = 1L, pairs = NULL, cpus = n_cpus()) {
   
+  norm <- 'percent'
   validate_args()
+  
   .Call(C_beta_div, BDIV_BHATTACHARYYA, counts, margin, norm, pairs, cpus, NULL)
 }
 
@@ -230,15 +239,16 @@ bhattacharyya <- function (counts, norm = 'percent', margin = 1L, pairs = NULL, 
 #' A standard ecological metric quantifying the dissimilarity between communities.
 #' 
 #' @inherit documentation
+#' 
 #' @family Abundance metrics
 #' @seealso `beta_div()`, `vignette('bdiv')`, `vignette('bdiv_guide')`
 #' 
[…]
