+++ man/mcintosh.Rd
@@ -8,9 +8,8 @@ mcintosh(counts, margin = 1L, cpus = n_cpus())
 }
 \arguments{
 \item{counts}{A numeric matrix of count data (samples \eqn{\times} features).
-Also supports \code{phyloseq}, \code{rbiom}, \code{SummarizedExperiment}, and
-\code{TreeSummarizedExperiment} objects. See \code{vignette('performance')} for
-optimizing large datasets.}
+Typically contains absolute abundances (integer counts), though
+proportions are also accepted.}
 
 \item{margin}{The margin containing samples. \code{1} if samples are rows,
 \code{2} if samples are columns. Ignored when \code{counts} is a special object
@@ -28,6 +27,7 @@ The McIntosh index is defined as:
 
 Where:
 \itemize{
+\item \eqn{n} : The number of features.
 \item \eqn{X_i} : Integer count of the \eqn{i}-th feature.
 \item \eqn{X_T} : Total of all counts.
 }
@@ -37,6 +37,23 @@ Where:
 \if{html}{\out{<div class="sourceCode r">}}\preformatted{(sum(x) - sqrt(sum(x^2))) / (sum(x) - sqrt(sum(x)))
 }\if{html}{\out{</div>}}
 }
+\section{Input Types}{
+
+
+The \code{counts} parameter is designed to accept a simple numeric matrix, but
+seamlessly supports objects from the following biological data packages:
+\itemize{
+\item \code{phyloseq}
+\item \code{rbiom}
+\item \code{SummarizedExperiment}
+\item \code{TreeSummarizedExperiment}
+}
+
+For large datasets, standard matrix operations may be slow. See
+\code{vignette('performance')} for details on using optimized formats
+(e.g. sparse matrices) and parallel processing.
+}
+
 \examples{
     mcintosh(ex_counts)
 }

[…]
