+++ reference/psym_chisq.md
@@ -6,39 +6,16 @@ distributions.
 ## Usage
 
 ``` r
-psym_chisq(
-  counts,
-  norm = "percent",
-  margin = 1L,
-  pairs = NULL,
-  cpus = n_cpus()
-)
+psym_chisq(counts, margin = 1L, pairs = NULL, cpus = n_cpus())
 ```
 
 ## Arguments
 
 - counts:
 
-  A numeric matrix of count data (samples \\\times\\ features). Also
-  supports `phyloseq`, `rbiom`, `SummarizedExperiment`, and
-  `TreeSummarizedExperiment` objects. See
-  [`vignette('performance')`](https://cmmr.github.io/ecodive/articles/performance.md)
-  for optimizing large datasets.
-
-- norm:
-
-  Normalize the incoming counts. Options are:
-
-  - `'percent'`: Relative abundance (sample abundances sum to 1).
-
-  - `'binary'`: Unweighted presence/absence (each count is either 0 or
-    1).
-
-  - `'clr'`: Centered log ratio.
-
-  - `'none'`: No transformation.
-
-  Default: `'percent'`, which is the expected input for these formulas.
+  A numeric matrix of count data (samples \\\times\\ features).
+  Typically contains absolute abundances (integer counts), though
+  proportions are also accepted.
 
 - margin:
 
@@ -72,7 +49,28 @@ Where:
 
 Base R Equivalent:
 
-    2 * sum((x - y)^2 / (x + y))
+    p <- x / sum(x)
+    q <- y / sum(y)
+    2 * sum((p - q)^2 / (p + q))
+
+## Input Types
+
+The `counts` parameter is designed to accept a simple numeric matrix,
+but seamlessly supports objects from the following biological data
+packages:
+
+- `phyloseq`
+
+- `rbiom`
+
+- `SummarizedExperiment`
+
+- `TreeSummarizedExperiment`
+
+For large datasets, standard matrix operations may be slow. See
+[`vignette('performance')`](https://cmmr.github.io/ecodive/articles/performance.md)
+for details on using optimized formats (e.g. sparse matrices) and
+parallel processing.
 
 ## References
 
