### 2026-02-04 4801fc4
Fix for missing `rarefy()` documentation
