### 2026-02-05 6946da4
implement robust zero handling for CLR transformations (#14)
