+++ R/validate.r
@@ -12,8 +12,8 @@ validate_args <- function () {
   env  <- parent.frame()
   args <- ls(env)
   
-  # move counts, pseudocount, and margin to head of the line
-  args <- unique(c(intersect(c('counts', 'pseudocount', 'margin'), args), sort(args)))
+  # move counts, norm, pseudocount, and margin to head of the line
+  args <- unique(c(intersect(c('counts', 'norm', 'pseudocount', 'margin'), args), sort(args)))
   
   for (arg in args)
     do.call(paste0('validate_', arg), list(env))
@@ -281,21 +281,95 @@ validate_power <- function (env = parent.frame()) {
 
 
 validate_pseudocount <- function (env = parent.frame()) {
-  tryCatch(
-    with(env, {
+  with(env, {
+    
+    if (!identical(norm, NORM_CLR)) {
+      pseudocount <- NULL
+    }
+    
+    else {
+      
+      mtx_pkg <- get_matrix_package(counts)
+      
+      val_range <- switch(
+        mtx_pkg,
+        'base'   = range(counts),
+        'slam'   = range(counts$v),
+        'Matrix' = range(counts@x) )
+      
+      if (length(val_range) == 2 && !all(is.finite(val_range)))
+        stop('`counts` contains non-finite values; cannot perform CLR normalization.')
+      
+      min_val <- val_range[1]
+      max_val <- val_range[2]
+      
+      if (min_val < 0)
+        stop('`counts` contains negative values; cannot perform CLR normalization.')
+      
+      if (max_val == 0)
+        stop('`counts` contains only zeros; cannot perform CLR normalization.')
+      
+      if (min_val == max_val)
+        stop('`counts` contains a single unique value; cannot perform CLR normalization.')
+      
       
       if (!is.null(pseudocount)) {
-        pseudocount <- as.double(pseudocount)
-        stopifnot(length(pseudocount) == 1)
-        stopifnot(!is.na(pseudocount))
-        stopifnot(pseudocount >= 0)
+        
+        tryCatch(
+          expr = {
+            stopifnot(is.numeric(pseudocount))
+            pseudocount <- as.double(pseudocount)
+            stopifnot(length(pseudocount) == 1)
+            stopifnot(!is.na(pseudocount))
+            stopifnot(pseudocount > 0)
+          },
+          error = function (e) {
+            stop('`pseudocount` must be a single positive number.')
+        })
+        
+      }
+      else {
+        
+        has_zeros <- min_val == 0
+        
+        # Check for zeros in sparse matrices
+        if (!has_zeros && mtx_pkg != 'base')
+          has_zeros <- switch(
+            mtx_pkg,
+            'slam'   = length(counts$v) < counts$nrow * counts$ncol,
+            'Matrix' = length(counts@x) < prod(dim(counts)) )
+        
+        
+        if (!has_zeros) {
+          
+          pseudocount <- NULL
+          
+        } else {
+          
+          # Calculate a "smart" default (half the smallest non-zero value)
+          # This is generally safer than '1' for proportional data.
+          pseudocount <- switch(
+            mtx_pkg,
+            'base'   = min(counts[counts > 0]),
+            'slam'   = min(counts$v[counts$v > 0]),
+            'Matrix' = min(counts@x[counts@x > 0]) )
+          
+          pseudocount <- pseudocount / 2
+          
+          warning(
+            call. = FALSE, 
+            paste0(
+              "Zeros detected in data. Using pseudocount = ", format(pseudocount, digits = 3), "\n",
[…]
