+++ tests/testthat/test-validate.r
@@ -59,17 +59,6 @@ test_that("validation", {
 
 
 
-
-  # validate_pseudocount() ====
-
-  env$pseudocount <- NULL
-  expect_silent(validate_pseudocount(env))
-
-  env$pseudocount <- 1L
-  expect_silent(validate_pseudocount(env))
-
-
-
   
   # validate_norm() ====
 
@@ -170,6 +159,137 @@ test_that("validation", {
   env$tree   <- NULL
   env$counts <- convert_to_SE(hmp50)
   expect_silent(validate_counts(env))
+  
+  
+  
+  
+  # validate_pseudocount() ====
+  
+  # Mock constants used in validate.r
+  NORM_PERCENT <- 1L
+  NORM_CLR     <- 2L
+  
+  # 1. Ignore when norm is NOT CLR
+  # ----------------------------------------------------------------
+  env$norm        <- NORM_PERCENT
+  env$pseudocount <- 999
+  env$counts      <- matrix(1:4, 2, 2)
+  expect_silent(validate_pseudocount(env))
+  expect_null(env$pseudocount)
+  
+  
+  # 2. Explicit Pseudocount (Valid)
+  # ----------------------------------------------------------------
+  env$norm        <- NORM_CLR
+  env$pseudocount <- 0.1
+  expect_silent(validate_pseudocount(env))
+  expect_equal(env$pseudocount, 0.1)
+  
+  
+  # 3. Explicit Pseudocount (Invalid inputs)
+  # ----------------------------------------------------------------
+  env$norm <- NORM_CLR
+  
+  env$pseudocount <- -1
+  expect_error(validate_pseudocount(env), "pseudocount` must be a single positive number")
+  
+  env$pseudocount <- 0
+  expect_error(validate_pseudocount(env), "pseudocount` must be a single positive number")
+  
+  env$pseudocount <- c(1, 2)
+  expect_error(validate_pseudocount(env), "pseudocount` must be a single positive number")
+  
+  env$pseudocount <- NA_real_
+  expect_error(validate_pseudocount(env), "pseudocount` must be a single positive number")
+  
+  env$pseudocount <- "string"
+  expect_error(validate_pseudocount(env), "pseudocount` must be a single positive number")
+  
+  
+  # 4. Data Validation: Negative / Infinite / NA values in counts
+  # ----------------------------------------------------------------
+  env$norm        <- NORM_CLR
+  env$pseudocount <- NULL
+  
+  env$counts <- matrix(c(1, 2, -1, 4), 2, 2)
+  expect_error(validate_pseudocount(env), "contains negative values")
+  
+  env$counts <- matrix(c(1, 2, Inf, 4), 2, 2)
+  expect_error(validate_pseudocount(env), "contains non-finite values")
+  
+  env$counts <- matrix(c(1, 2, NA, 4), 2, 2)
+  expect_error(validate_pseudocount(env), "contains non-finite values")
+  
+  
+  # 5. Auto-detection: Dense Matrix
+  # ----------------------------------------------------------------
+  env$norm <- NORM_CLR
+  
+  # Case: No zeros -> No change
+  env$counts      <- matrix(1:4, 2, 2)
+  env$pseudocount <- NULL
+  expect_silent(validate_pseudocount(env))
+  expect_null(env$pseudocount)
+  
+  # Case: Zeros present -> Warning + Default (min/2)
+  env$counts      <- matrix(c(0, 10, 20, 30), 2, 2) # min non-zero is 10
+  env$pseudocount <- NULL
[…]
