### 2026-02-05 b76bccb
Added `x` and `y` defs to "Base R Equivalent" documentation sections.
