+++ R/beta_div.r
@@ -231,8 +231,8 @@ aitchison <- function (counts, margin = 1L, pseudocount = NULL, pairs = NULL, cp
 #' 
 #' Base R Equivalent: 
 #' ```r
-#' p <- x / sum(x)
-#' q <- y / sum(y)
+#' x <- ex_counts[1,]; p <- x / sum(x)
+#' y <- ex_counts[2,]; q <- y / sum(y)
 #' -log(sum(sqrt(p * q)))
 #' ```
 #' 
@@ -271,6 +271,8 @@ bhattacharyya <- function (counts, margin = 1L, pairs = NULL, cpus = n_cpus()) {
 #' 
 #' Base R Equivalent: 
 #' ```r
+#' x <- ex_counts[1,]
+#' y <- ex_counts[2,]
 #' sum(abs(x-y)) / sum(x+y)
 #' ```
 #' 
@@ -307,6 +309,8 @@ bray <- function (counts, margin = 1L, norm = 'none', pseudocount = NULL, pairs 
 #' 
 #' Base R Equivalent: 
 #' ```r
+#' x <- ex_counts[1,]
+#' y <- ex_counts[2,]
 #' sum(abs(x-y) / (x+y))
 #' ```
 #' 
@@ -342,6 +346,8 @@ canberra <- function (counts, margin = 1L, norm = 'none', pseudocount = NULL, pa
 #' 
 #' Base R Equivalent: 
 #' ```r
+#' x <- ex_counts[1,]
+#' y <- ex_counts[2,]
 #' max(abs(x-y))
 #' ```
 #' 
@@ -377,6 +383,8 @@ chebyshev <- function (counts, margin = 1L, norm = 'none', pseudocount = NULL, p
 #' 
 #' Base R Equivalent: 
 #' ```r
+#' x <- ex_counts[1,]
+#' y <- ex_counts[2,]
 #' sqrt(sum(((x / sqrt(sum(x ^ 2))) - (y / sqrt(sum(y ^ 2))))^2))
 #' ```
 #' 
@@ -415,6 +423,8 @@ chord <- function (counts, margin = 1L, pairs = NULL, cpus = n_cpus()) {
 #' 
 #' Base R Equivalent: 
 #' ```r
+#' x <- ex_counts[1,]
+#' y <- ex_counts[2,]
 #' sqrt(sum((abs(x - y) / (x + y)) ^ 2))
 #' ```
 #' 
@@ -452,8 +462,8 @@ clark <- function (counts, margin = 1L, norm = 'none', pseudocount = NULL, pairs
 #' 
 #' Base R Equivalent: 
 #' ```r
-#' p <- x / sum(x)
-#' q <- y / sum(y)
+#' x <- ex_counts[1,]; p <- x / sum(x)
+#' y <- ex_counts[2,]; q <- y / sum(y)
 #' 2 * sum((p - q)^2 / (p + q)^2)
 #' ```
 #' 
@@ -492,6 +502,8 @@ divergence <- function (counts, margin = 1L, norm = 'none', pseudocount = NULL, 
 #' 
 #' Base R Equivalent: 
 #' ```r
+#' x <- ex_counts[1,]
+#' y <- ex_counts[2,]
 #' sqrt(sum((x-y)^2))
 #' ```
 #' 
@@ -529,8 +541,10 @@ euclidean <- function (counts, margin = 1L, norm = 'none', pseudocount = NULL, p
 #' 
 #' Base R Equivalent: 
 #' ```r
+#' x <- ex_counts[1,]
+#' y <- ex_counts[2,]
 #' r <- abs(x - y)
[…]
