+++ man/sorensen.Rd
@@ -38,7 +38,9 @@ Where:
 
 Base R Equivalent:
 
-\if{html}{\out{<div class="sourceCode r">}}\preformatted{2 * sum(x & y) / sum(x>0, y>0)
+\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- ex_counts[1,]
+y <- ex_counts[2,]
+2 * sum(x & y) / sum(x>0, y>0)
 }\if{html}{\out{</div>}}
 }
 \section{Input Types}{
