+++ reference/morisita.md
@@ -54,9 +54,11 @@ Where:
 
 Base R Equivalent:
 
-    simp_x <- sum(x * (x - 1)) / (sum(x) * (sum(x) - 1))
-    simp_y <- sum(y * (y - 1)) / (sum(y) * (sum(y) - 1))
-    1 - ((2 * sum(x * y)) / ((simp_x + simp_y) * sum(x) * sum(y)))
+    x <- ex_counts[1,]
+    y <- ex_counts[2,]
+    simpson_x <- sum(x * (x - 1)) / (sum(x) * (sum(x) - 1))
+    simpson_y <- sum(y * (y - 1)) / (sum(y) * (sum(y) - 1))
+    1 - ((2 * sum(x * y)) / ((simpson_x + simpson_y) * sum(x) * sum(y)))
 
 ## Input Types
 
