+++ reference/soergel.md
@@ -75,6 +75,8 @@ Where:
 
 Base R Equivalent:
 
+    x <- ex_counts[1,]
+    y <- ex_counts[2,]
     sum(abs(x - y)) / sum(pmax(x, y))
 
 ## Input Types
