+++ reference/hamming.md
@@ -48,6 +48,8 @@ Where:
 
 Base R Equivalent:
 
+    x <- ex_counts[1,]
+    y <- ex_counts[2,]
     sum(xor(x, y))
 
 ## Input Types
