+++ reference/canberra.md
@@ -76,6 +76,8 @@ Where:
 
 Base R Equivalent:
 
+    x <- ex_counts[1,]
+    y <- ex_counts[2,]
     sum(abs(x-y) / (x+y))
 
 ## Input Types
