+++ reference/sorensen.md
@@ -48,6 +48,8 @@ Where:
 
 Base R Equivalent:
 
+    x <- ex_counts[1,]
+    y <- ex_counts[2,]
     2 * sum(x & y) / sum(x>0, y>0)
 
 ## Input Types
