### 2026-02-05 #9777 @dansmith01
Thank you for your thorough feedback @CosteaPaul; it was very useful for improving the package. I have addressed the issues you raised in the latest updates:

* **Default Normalization**: Changed the default for `norm` from `"percent"` to `"none"` to ensure functions do not modify data unless explicitly requested.

* **CLR/Aitchison Zero Handling**: To encourage informed parameter selection, the package now issues a warning whenever it automatically calculates a pseudocount, specifying both the formula applied and the resulting value.

* **Rarefaction Predictability**: Simplified the `rarefy()` function by removing the `n_samples` and fractional `depth` parameters and refining the `drop` behavior for samples and features.

I’ve left the corresponding issues open on GitHub so you can review the changes and close them if they meet your expectations, or provide further feedback if needed. Thanks again for your time and the detailed review!
