### 2026-02-06 f24193d
Autodetect when `counts` is already proportions.
