### 2026-02-06 5d24df1
centralized `run_parallel()` C function
