### 2026-02-06 2782c32
fix for `error: expected expression`
