### 2026-02-06 c0fee0b
benchmarking code updated to remove percent normalization
