+++ articles/benchmark.md
@@ -23,7 +23,7 @@ implementations.
 | [GUniFrac](https://doi.org/10.32614/CRAN.package.GUniFrac)         | *none*               | ***Serial C***          |
 | [labdsv](https://doi.org/10.32614/CRAN.package.labdsv)             | ***Serial FORTRAN*** | *none*                  |
 | [parallelDist](https://doi.org/10.32614/CRAN.package.parallelDist) | ***Parallel C++***   | *none*                  |
-| [philentropy](https://doi.org/10.32614/CRAN.package.philentropy)   | ***Serial C++***     | *none*                  |
+| [philentropy](https://doi.org/10.32614/CRAN.package.philentropy)   | ***Parallel C++***   | *none*                  |
 | [phyloregion](https://doi.org/10.32614/CRAN.package.phyloregion)   | vegan                | ***Serial R***          |
 | [phyloseq](https://doi.org/doi:10.18129/B9.bioc.phyloseq)          | vegan                | ***Parallel R***        |
 | [picante](https://doi.org/10.32614/CRAN.package.picante)           | vegan                | ***Serial R***          |
@@ -89,16 +89,16 @@ data.frame(ver = sapply(FUN = packageDescription, fields = 'Version', c(
   'abdiv', 'adiv', 'ecodist', 'ecodive', 'entropart', 'GUniFrac', 
   'ampvis2', 'labdsv', 'parallelDist', 'philentropy', 
   'phyloregion', 'phyloseq', 'picante', 'tabula', 'vegan' )))
-#>                 ver
+#>                 ver                   ver
 #> bench         1.1.4
-#> dplyr         1.1.4
-#> ggplot2       4.0.1
+#> dplyr         1.2.0
+#> ggplot2       4.0.2
 #> ggrepel       0.9.6
-#> rbiom         2.2.1
+#> rbiom         3.0.1
 #> abdiv         0.2.0
 #> adiv          2.2.1
 #> ecodist       2.1.3
-#> ecodive       2.2.0
+#> ecodive       2.2.5
 #> entropart    1.6-16
 #> GUniFrac        1.9
 #> ampvis2      2.8.11
@@ -146,27 +146,35 @@ cleanup <- function (x) {
 }
 
 
-# HMP50 dataset has 50 samples; convert to relative abundances
-hmp50       <- rbiom::hmp50
-hmp50_phy   <- rbiom::convert_to_phyloseq(hmp50)
-hmp50_mtx   <- t(apply(as.matrix(hmp50), 2L, function (x) x / sum(x)))
-hmp50_mtx_t <- t(hmp50_mtx)
-hmp50_dgC   <- as(hmp50_mtx, 'dgCMatrix')
-hmp50_dgC_t <- t(hmp50_dgC)
-hmp50_tree  <- hmp50$tree
+# HMP50 dataset has 50 samples
+hmp50        <- rbiom::hmp50
+hmp50_phy    <- rbiom::convert_to_phyloseq(hmp50)
+hmp50_mtx_t  <- as.matrix(hmp50)
+hmp50_mtx    <- t(hmp50_mtx_t)
+hmp50_mtx_pt <- apply(hmp50_mtx, 1L, function (x) x / sum(x))
+hmp50_mtx_p  <- t(hmp50_mtx_pt)
+hmp50_dgC    <- as(hmp50_mtx,    'dgCMatrix')
+hmp50_dgC_t  <- as(hmp50_mtx_t,  'dgCMatrix')
+hmp50_dgC_pt <- as(hmp50_mtx_pt, 'dgCMatrix')
[…]
