### 2026-02-07 b6b6114
ensure `pseudocount` is always a double when passed to C
