### 2026-02-07 #9777 @dansmith01
Hi @cstawitz, I have addressed all the open issues in the ecodive repository (https://github.com/cmmr/ecodive/issues) and confirmed that the relevant commits are tagged in each response. Since the review process is currently waiting on your final approval, could you please close those issues and leave a comment here if you are satisfied with the changes? Tagging @goldingn for visibility. Thanks!
