+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: ecodive
 Type: Package
 Title: Parallel and Memory-Efficient Ecological Diversity Metrics
-Version: 2.2.5
+Version: 2.2.6
 Authors@R: c(
     person(
       "Daniel P.", "Smith",
