### 2026-02-23 #9777 @dansmith01
Hi @cstawitz,

I wanted to check in and see if you’ve had a chance to review the recent updates to the manuscript and documentation. Since we last spoke, I've addressed the items we discussed, and I'm happy to provide any further clarification if needed.

@goldingn — as we have sign-off from the other reviewer and I've addressed the feedback from @cstawitz, I was wondering if there are any further steps you'd like me to take to move this toward a conclusion, or if we should continue to wait for a final look here?
