### 2026-02-26 #9777 @cstawitz
Thank you @dansmith01 for the changes to the manuscript, adding citations to the vignette, and the expansion of the documentation. It's much easier to follow and more descriptive with the additions. I'm satisfied with the changes and support publishing the manuscript, well done.
