### 2026-03-03 #9777 @dansmith01
Hi @goldingn,

It looks like both @cstawitz and @CosteaPaul have now completed their checklists and are satisfied with the revisions.

I’d like to extend a sincere thank you to @cstawitz and @CosteaPaul for their time and effort in reviewing ecodive. Your feedback and suggestions have been incredibly helpful and have significantly improved the package and its documentation.

@goldingn, please let me know if there are any further steps I need to take on my end to move forward with the submission!
