### 2026-03-06 #9777 @goldingn
Hi @dansmith01 I'm so sorry for the long absence and lack of communication. I'll be back on deck on Monday, and I'll prioritise getting this one finalised.

@CosteaPaul thanks for giving the thumbs up. Sorry to be a pain, but could you please create a checklist (type `@editorialbot generate my checklist`) and check off the items (if you are happy with them)? It's important for our reviews that you've been through all of the criteria.

@cstawitz I see you are happy with the MS, can you please confirm (or ideally check off on the list) that you are now happy with those last unchecked items?
