### 2026-03-09 #9777 @CosteaPaul
@goldingn , did the checklist stuff, though you can be 100% sure that i did not read carefully.

I do like the transparency of doing reviews this way and i'm generally happy with the concept. But since this is purely pro bono i find the hassle quite annoying. I have to do unnecessary stuff like this for work too, but at least they pay me ;)
