### 2026-03-23 #9777 @goldingn
- [ ] Check off any remaining boxes [here](https://github.com/openjournals/joss-reviews/issues/9777#issuecomment-3768712962)
