### 2026-03-23 #9777 @goldingn
Hi @dansmith01, thanks for your patience. There are a few checkboxes missing from one review, but I can assess those when I go through the paper as part of my next steps.

Before I can do that, I need you to create an archive of the software (v2.2 or we can update that too if needed), making sure the title, author list, and license for the archive are the same as in the JOSS paper, as outlined here: https://github.com/openjournals/joss-reviews/issues/9777#issuecomment-4108831272
