### 2026-03-23 #9777 @dansmith01
Hi @goldingn,

I have just created a new release on GitHub that incorporates all of the revisions requested by the reviewers. Could you please bump the version to `2.2.6`?

The Zenodo DOI for this version is `10.5281/zenodo.19190015` (URL: https://doi.org/10.5281/zenodo.19190015).

I have also gone through the Post-Review Checklist and believe I have everything checked off. Please let me know if there are any further steps required to finalize the submission.
