+++ R/data.r
@@ -8,7 +8,6 @@
 #' 
 #' @format A matrix of 4 samples (columns) x 6 genera (rows).
 #' @source Derived from The Human Microbiome Project dataset.
-#' <https://commonfund.nih.gov/hmp>
 "ex_counts"
 
 
