+++ reference/psym_chisq.md
@@ -82,7 +82,7 @@ Mathematical Models and Methods in Applied Sciences*, 1(4), 300–307.
 
 [`beta_div()`](https://cmmr.github.io/ecodive/reference/beta_div.md),
 [`vignette('bdiv')`](https://cmmr.github.io/ecodive/articles/bdiv.md),
-[`vignette('bdiv_guide')`](https://cmmr.github.io/ecodive/articles/bdiv_guide.md)
+`vignette('bdiv_guide')`
 
 Other Abundance metrics:
 [`aitchison()`](https://cmmr.github.io/ecodive/reference/aitchison.md),
