+++ man/aitchison.Rd
@@ -92,7 +92,7 @@ For large datasets, standard matrix operations may be slow. See
     aitchison(ex_counts, pseudocount = 1)
 }
 \references{
-Aitchison, J. (1986). The statistical analysis of compositional data. Chapman and Hall. \doi{10.1007/978-94-009-4109-3}
+Aitchison, J. (1986). The statistical analysis of compositional data. Chapman and Hall. \doi{10.1002/bimj.4710300705}
 
 Aitchison, J. (1982). The statistical analysis of compositional data. \emph{Journal of the Royal Statistical Society: Series B (Methodological)}, 44(2), 139-160. \doi{10.1111/j.2517-6161.1982.tb01195.x}
 
