### 2026-04-14 78ea1a1
bumped version in `NEWS.md`
