### 2026-04-27 273fc15
`read_tree` now properly protects against garbage collections.
