+++ src/read_tree.c
@@ -10,11 +10,13 @@ static int         underscores;
 static size_t n_leafs;
 static size_t n_edges;
 
-// Output arrays
+// Output arrays (R-managed handles)
 static SEXP    sexp_node_label_vec;
 static SEXP    sexp_leaf_label_vec;
-static int    *edge_mtx;
-static double *edge_length_vec;
+
+// Native C arrays to avoid R garbage collection relocation issues
+static int    *c_edge_mtx;
+static double *c_edge_length_vec;
 
 // Track next open indices in output arrays
 static size_t next_edge_index;
@@ -22,18 +24,20 @@ static size_t next_node_index;
 static size_t next_leaf_index;
 
 
-
-
 static SEXP extract_name(size_t x1, size_t x2) {
+  
+  if (x1 > x2) return mkChar(""); // Safety net for empty ranges
   
   char quoted = tree_str[x1] == '\'' && tree_str[x2] == '\'';
   
   // Quoted Name ==> Strip off quote marks
   if (quoted) {
+    if (x2 - x1 < 2) return mkChar(""); // e.g. ''
     x1++;
     x2--;
   }
   
+  // Standard calloc (heap), completely invisible to R's memory tracker
   char* node_name_str = calloc(x2 - x1 + 2, sizeof(char));
   strncpy(node_name_str, tree_str + x1, x2 - x1 + 1);
   node_name_str[x2 - x1 + 1] = '\0';
@@ -46,12 +50,12 @@ static SEXP extract_name(size_t x1, size_t x2) {
   }
   
   SEXP sexp_node_name = mkChar(node_name_str);
+  
+  // Free immediately! Prevents massive numbers of tiny heap allocations from piling up
   free(node_name_str);
   
   return sexp_node_name;
 }
-
-
 
 
 static void recurse_tree(size_t x1, size_t x2, size_t parent) {
@@ -75,9 +79,9 @@ static void recurse_tree(size_t x1, size_t x2, size_t parent) {
     // Text after the colon is the branch length
     if (tree_str[i] == ':') {
       
-      if (next_edge_index > 0) {
+      if (next_edge_index > 0 && next_edge_index <= n_edges) {
         char *junk_ptr;
-        edge_length_vec[next_edge_index - 1] = strtod(tree_str + i + 1, &junk_ptr);
+        c_edge_length_vec[next_edge_index - 1] = strtod(tree_str + i + 1, &junk_ptr);
       }
       
       x2 = i - 1;
@@ -89,10 +93,10 @@ static void recurse_tree(size_t x1, size_t x2, size_t parent) {
       if (i < x2)
         SET_STRING_ELT(sexp_node_label_vec, next_node_index, extract_name(i + 1, x2));
       
-      if (next_edge_index > 0) {
+      if (next_edge_index > 0 && next_edge_index <= n_edges) {
         size_t mtx_row_i = next_edge_index - 1;
-        edge_mtx[mtx_row_i + 0]       = parent + n_leafs;
-        edge_mtx[mtx_row_i + n_edges] = next_node_index + n_leafs + 1;
+        c_edge_mtx[mtx_row_i + 0]       = parent + n_leafs;
+        c_edge_mtx[mtx_row_i + n_edges] = next_node_index + n_leafs + 1;
       }
       next_node_index++;
       next_edge_index++;
[…]
