### 2026-04-27 #9777 @dansmith01
Hi @goldingn, I’m checking in again to see if we can move this toward final acceptance.

It has been about a month since the post-review tasks were completed. As a reminder, the final release is `v2.2.6` and the Zenodo archive is `10.5281/zenodo.19190015`.

Since both reviewers have given their approval, please let me know if there is anything else you need from my side, or if you are all set to proceed with the final editorial checks. Thanks!
