+++ tests/tinytests/test-alpha_div.r
@@ -1,4 +1,4 @@
-test_that("alpha diversity", {
+#test_that("alpha diversity", {
   
   
   
@@ -20,8 +20,8 @@ test_that("alpha diversity", {
   # ACE ====
   
   expect_equal( # vegan::estimateR(counts)['S.ACE',]
-    object   = ace(counts, cutoff = 5), 
-    expected = c(A = 3, B = 3, C = 3, D = 2) )
+    current = ace(counts, cutoff = 5), 
+    target  = c(A = 3, B = 3, C = 3, D = 2) )
   
   expect_error(ace(counts, cutoff = 'A'))  # stopifnot(is.numeric(cutoff))
   expect_error(ace(counts, cutoff = NULL)) # stopifnot(length(cutoff) == 1)
@@ -34,8 +34,8 @@ test_that("alpha diversity", {
   # Berger ====
   
   expect_equal( # apply(counts, 1, tabula::index_berger)
-    object   = berger(counts), 
-    expected = c(A = 0.461538461538462, B = 0.444444444444444, 
+    current = berger(counts), 
+    target  = c(A = 0.461538461538462, B = 0.444444444444444, 
                  C = 0.428571428571429, D = 0.666666666666667 ))
   
   
@@ -43,8 +43,8 @@ test_that("alpha diversity", {
   # Brillouin ====
   
   expect_equal( # apply(counts, 1, tabula::index_brillouin)
-    object   = brillouin(counts), 
-    expected = c(A = 0.807097978290102, B = 0.900881045540206, 
+    current = brillouin(counts), 
+    target  = c(A = 0.807097978290102, B = 0.900881045540206, 
                  C = 0.91741230069276,  D = 0.533824471198889 ))
   
   
@@ -52,8 +52,8 @@ test_that("alpha diversity", {
   # Chao1 ====
   
   expect_equal(
-    object   = chao1(counts), 
-    expected = c(A = 3, B = NaN, C = NaN,  D = NaN) )
+    current = chao1(counts), 
+    target  = c(A = 3, B = NaN, C = NaN,  D = NaN) )
   
   expect_equal(chao1(t(1:10)), 10.5)
   expect_equal(chao1(t(rep(1:10, 2))), 21)
@@ -63,16 +63,16 @@ test_that("alpha diversity", {
   # Faith's Phylogenetic Diversity ====
  
   expect_equal( # apply(counts, 1L, abdiv::faith_pd, tree)
-    object   = faith(counts, tree), 
-    expected = c(A = 2.319, B = 2.191, C = 2.191, D = 1.759) )
+    current = faith(counts, tree), 
+    target  = c(A = 2.319, B = 2.191, C = 2.191, D = 1.759) )
   
   
   
   # Fisher ====
   
   expect_equal( # vegan::fisher.alpha(counts)
-    object   = fisher(counts, digits = 5), 
-    expected = c(A = 1.22255, B = 1.02800, C = 0.95777, D = 0.61979) )
+    current = fisher(counts, digits = 5), 
+    target  = c(A = 1.22255, B = 1.02800, C = 0.95777, D = 0.61979) )
   
   expect_error(fisher(counts, digits = 'A'))  # stopifnot(is.numeric(digits))
   expect_error(fisher(counts, digits = NULL)) # stopifnot(length(digits) == 1)
@@ -86,8 +86,8 @@ test_that("alpha diversity", {
   # Inverse Simpson ====
  
   expect_equal( # vegan::diversity(counts, 'invsimpson')
-    object   = inv_simpson(counts), 
-    expected = c(A = 2.6, B = 2.84210526315789, 
+    current = inv_simpson(counts), 
+    target  = c(A = 2.6, B = 2.84210526315789, 
                  C = 2.84516129032258, D = 1.8 ))
   
   
[…]
