### 2026-04-27 69f5d20
revert to 2.2.6 `read_tree` memory management
