+++ src/read_tree.c
@@ -11,12 +11,12 @@ static size_t n_leafs;
 static size_t n_edges;
 
 // Output arrays (R-managed handles)
-static SEXP    sexp_node_label_vec;
-static SEXP    sexp_leaf_label_vec;
+static SEXP sexp_node_label_vec;
+static SEXP sexp_leaf_label_vec;
 
 // Native C arrays to avoid R garbage collection relocation issues
-static int    *c_edge_mtx;
-static double *c_edge_length_vec;
+static int    *edge_mtx;
+static double *edge_length_vec;
 
 // Track next open indices in output arrays
 static size_t next_edge_index;
@@ -81,7 +81,7 @@ static void recurse_tree(size_t x1, size_t x2, size_t parent) {
       
       if (next_edge_index > 0 && next_edge_index <= n_edges) {
         char *junk_ptr;
-        c_edge_length_vec[next_edge_index - 1] = strtod(tree_str + i + 1, &junk_ptr);
+        edge_length_vec[next_edge_index - 1] = strtod(tree_str + i + 1, &junk_ptr);
       }
       
       x2 = i - 1;
@@ -95,8 +95,8 @@ static void recurse_tree(size_t x1, size_t x2, size_t parent) {
       
       if (next_edge_index > 0 && next_edge_index <= n_edges) {
         size_t mtx_row_i = next_edge_index - 1;
-        c_edge_mtx[mtx_row_i + 0]       = parent + n_leafs;
-        c_edge_mtx[mtx_row_i + n_edges] = next_node_index + n_leafs + 1;
+        edge_mtx[mtx_row_i + 0]       = parent + n_leafs;
+        edge_mtx[mtx_row_i + n_edges] = next_node_index + n_leafs + 1;
       }
       next_node_index++;
       next_edge_index++;
@@ -118,8 +118,8 @@ static void recurse_tree(size_t x1, size_t x2, size_t parent) {
     
     if (next_edge_index > 0 && next_edge_index <= n_edges) {
       size_t mtx_row_i = next_edge_index - 1;
-      c_edge_mtx[mtx_row_i + 0]       = parent + n_leafs;
-      c_edge_mtx[mtx_row_i + n_edges] = next_leaf_index + 1;
+      edge_mtx[mtx_row_i + 0]       = parent + n_leafs;
+      edge_mtx[mtx_row_i + n_edges] = next_leaf_index + 1;
     }
     next_leaf_index++;
     next_edge_index++;
@@ -193,8 +193,8 @@ SEXP C_read_tree(SEXP sexp_tree_str, SEXP sexp_underscores) {
   SEXP sexp_result          = PROTECT(allocVector(VECSXP, 5));
   SEXP sexp_edge_mtx        = PROTECT(allocMatrix(INTSXP,  n_edges, 2));
   SEXP sexp_edge_length_vec = PROTECT(allocVector(REALSXP, n_edges));
-  sexp_node_label_vec  = PROTECT(allocVector(STRSXP,  n_nodes));
-  sexp_leaf_label_vec  = PROTECT(allocVector(STRSXP,  n_leafs));
+       sexp_node_label_vec  = PROTECT(allocVector(STRSXP,  n_nodes));
+       sexp_leaf_label_vec  = PROTECT(allocVector(STRSXP,  n_leafs));
   
   // Initialize strings to ""
   for (size_t i = 0; i < n_nodes; i++) SET_STRING_ELT(sexp_node_label_vec, i, R_BlankString);
@@ -206,30 +206,21 @@ SEXP C_read_tree(SEXP sexp_tree_str, SEXP sexp_underscores) {
   SET_VECTOR_ELT(sexp_result, 3, sexp_edge_length_vec);
[…]
