### 2026-05-08 6462c93
Explicitly state six-core benchmarking in JOSS paper.
