+++ paper/paper.md
@@ -67,7 +67,7 @@ To address this, `ecodive` maintains the standard R interface (samples-as-rows) 
 
 `ecodive` has demonstrated immediate utility in high-dimensional microbiome studies. The core C algorithms in `ecodive` were originally developed for and deployed in the `rbiom` package [@rbiom]. As part of `rbiom`, these optimized metrics have already been utilized in diverse microbial ecology studies, including research on preterm infant microbiomes [@AhearnFord2025], dietary interventions [@DiMattia2025], and relationship satisfaction [@Cheng2023]. `ecodive` extracts these proven, high-performance components into a standalone, lightweight library to make them accessible to the broader R ecosystem without `rbiom`'s specific visualization and data structure dependencies.
 
-In [benchmarks](https://cmmr.github.io/ecodive/articles/benchmark.html) comparing 15 ecological R packages, `ecodive` consistently ranked as the fastest and most memory-efficient solution:
+In [benchmarks](https://cmmr.github.io/ecodive/articles/benchmark.html) comparing 15 ecological R packages, `ecodive` consistently ranked as the fastest and most memory-efficient solution. To reflect realistic modern research conditions, these tests were conducted on a workstation with a 6-core Intel i5-9600K (3.70GHz) CPU and 64GB RAM running Windows 11. While `ecodive`, `parallelDist`, `philentropy`, `phyloseq`, and `ampvis2` leveraged all six available CPU cores, the results show that parallelization is not a "silver bullet" for performance. In several metrics, specialized serial implementations outperformed parallel-capable packages - for example, the serial `phyloregion` was significantly faster than the parallelized `ampvis2` for UniFrac calculations - indicating that efficient data handling and core algorithms are as critical as raw multi-threading.
 
 * **Speed:** For the widely-used Unweighted UniFrac metric ($N=50$), `ecodive` completed calculations in 6.4ms, compared to 2.5s for `picante` (396x faster) and 297ms for `phyloseq` (46x faster).
 * **Scalability:** For standard Bray-Curtis dissimilarity ($N=1006$), `ecodive` processed the matrix in ~20ms, whereas `vegan` required 1.68s.
