### 2026-05-14 315f9eb
paper and benchmarks updates
