+++ paper/paper.md
@@ -33,11 +33,11 @@ bibliography: paper.bib
 
 # Summary
 
-Understanding the complexity of biological communities - whether bacteria in the human gut, trees in a forest, or plankton in the ocean - is a central goal of ecology. Researchers quantify this complexity using "diversity metrics," which describe the variety of species within a single site (alpha diversity) or the differences in composition between two sites (beta diversity). `ecodive` is an R package designed to calculate these metrics efficiently. It bridges the gap between complex ecological theory and practical data analysis, providing researchers with a unified toolset to process large-scale datasets that were previously computationally prohibitive. By leveraging parallel processing and optimized memory management, `ecodive` enables rapid, high-throughput analysis of microbial and macro-ecological communities.
+Quantifying the biological complexity of communities - ranging from human gut microbiomes to forest ecosystems - is fundamental to modern ecology. Researchers use diversity metrics to describe the variety within a site (alpha diversity) and the compositional differences between sites (beta diversity). `ecodive` is an R package built to compute these metrics with high efficiency. It provides a comprehensive, unified toolset that allows researchers to analyze large-scale datasets that were previously computationally prohibitive. By integrating optimized C-based algorithms with a parallel processing engine, `ecodive` enables rapid, high-throughput ecological analysis through a lightweight, portable framework.
 
 # Statement of Need
 
-A primary challenge in modern ecological analysis is the management of high-dimensional data. As sequencing technologies improve, datasets are growing to include thousands of samples and tens of thousands of unique taxa. Beta diversity calculations, which involve comparing every sample to every other sample, exhibit $O(n^2)$ complexity. This quadratic scaling creates a significant bottleneck; a dataset that doubles in size requires four times the processing power, often overwhelming standard desktop computers.
+A primary challenge in modern ecological analysis is the management of high-dimensional data. As sequencing technologies improve, datasets are growing to include thousands of samples and tens of thousands of unique taxa. Beta diversity calculations, which involve comparing every sample to every other sample, exhibit $O(n^2)$ complexity. This quadratic scaling creates a significant bottleneck; doubling a dataset's size quadruples the processing time and memory requirements, frequently exceeding the capacity of standard research workstations.
 
 Furthermore, the software landscape for ecological metrics is fragmented. A researcher needing to calculate a specific set of indices - for example, Faith's Phylogenetic Diversity, Bray-Curtis dissimilarity, and UniFrac distances - often must install and manage multiple R packages (`picante`, `vegan`, `GUniFrac`), each with different dependencies, input formats, and performance limitations.
 
@@ -45,11 +45,43 @@ Furthermore, the software landscape for ecological metrics is fragmented. A rese
 
 # State of the Field
 
-Several R packages exist for diversity analysis, but `ecodive` offers a unique contribution through its scope and performance. The standard package for community ecology, `vegan` [@vegan], provides excellent implementations of non-phylogenetic metrics (e.g., Bray-Curtis) but lacks phylogenetic awareness (e.g., UniFrac). Conversely, packages like `picante` [@picante] and `GUniFrac` [@GUniFrac] specialize in phylogenetic metrics but do not offer a comprehensive suite of general-purpose indices. The `phyloseq` [@phyloseq] package wraps many of these tools but relies on their underlying, often serial, implementations.
[…]
