+++ paper/paper.bib
@@ -1,21 +1,110 @@
-@manual{GUniFrac,
-  title  = {GUniFrac: Generalized UniFrac Distances, Distance-Based
+@misc{abdiv,
+  title     = {abdiv: Alpha and Beta Diversity Measures},
+  author    = {Bittinger, Kyle},
+  year      = {2020},
+  journal   = {CRAN: Contributed Packages},
+  publisher = {The R Foundation},
+  doi       = {10.32614/cran.package.abdiv}
+}
+
+@article{adiv,
+  title   = {adiv: An r package to analyse biodiversity in ecology},
+  author  = {Sandrine Pavoine},
+  year    = {2020},
+  journal = {Methods in Ecology and Evolution},
+  volume  = {11},
+  pages   = {1106-1112},
+  doi     = {10.1111/2041-210X.13430}
+}
+
+@article{ampvis2,
+  title     = {ampvis2: an R package to analyse and visualise 16S rRNA amplicon data},
+  author    = {Andersen, Kasper S. and Kirkegaard, Rasmus H. and Karst, Søren M. and Albertsen, Mads},
+  year      = {2018},
+  publisher = {openRxiv},
+  doi       = {10.1101/299537}
+}
+
+@article{ecodist,
+  title   = {The ecodist package for dissimilarity-based analysis of ecological data},
+  author  = {Sarah C. Goslee and Dean L. Urban},
+  year    = {2007},
+  journal = {Journal of Statistical Software},
+  volume  = {22},
+  issue   = {7},
+  pages   = {1-19},
+  doi     = {10.18637/jss.v022.i07}
+}
+
+@article{entropart,
+  title   = {entropart: An R Package to Measure and Partition Diversity},
+  author  = {Eric Marcon and Bruno Herault},
+  year    = {2015},
+  journal = {Journal of Statistical Software},
+  volume  = {67},
+  number  = {8},
+  pages   = {1-26},
+  doi     = {10.18637/jss.v067.i08}
+}
+
+@misc{GUniFrac,
+  title     = {GUniFrac: Generalized UniFrac Distances, Distance-Based
     Multivariate Methods and Feature-Based Univariate Methods for
     Microbiome Data Analysis},
-  author = {Jun Chen and Xianyang Zhang and Lu Yang and Lujun Zhang},
-  year   = {2023},
-  doi    = {10.32614/CRAN.package.GUniFrac}
+  author    = {Jun Chen and Xianyang Zhang and Lu Yang and Lujun Zhang},
+  year      = {2023},
+  journal   = {CRAN: Contributed Packages},
+  publisher = {The R Foundation}, 
+  doi       = {10.32614/CRAN.package.GUniFrac}
+}
+
+@misc{labdsv,
+  title     = {labdsv: Ordination and Multivariate Analysis for Ecology},
+  author    = {Roberts, David W.},
+  year      = {2005},
+  journal   = {CRAN: Contributed Packages},
+  publisher = {The R Foundation},
+  doi       = {10.32614/cran.package.labdsv}
+}
+
+@misc{OmicFlow, 
+  title     = {OmicFlow: Fast and Efficient (Automated) Analysis of Sparse Omics Data},
+  author    = {Gusinac, Alem and Ederveen, Thomas and Boleij, Annemarie},
+  year      = {2025},
+  journal   = {CRAN: Contributed Packages},
+  publisher = {The R Foundation},
+  doi       = {10.32614/cran.package.omicflow}
+}
+
+@misc{parallelDist,
+  title     = {parallelDist: Parallel Distance Matrix Computation using Multiple Threads},
+  author    = {Eckert, Alexander},
+  year      = {2017},
+  journal   = {CRAN: Contributed Packages},
[…]
