### 2026-05-14 e95b594
table column widths
