### 2026-05-14 9b200dd
adjust table widths
