### 2026-05-14 7851e2c
adjust table column widths
