+++ paper/paper.md
@@ -45,7 +45,7 @@ Furthermore, the software landscape for ecological metrics is fragmented. A rese
 
 # State of the Field
 
-While several R packages support diversity analysis, ecodive offers a unique scholarly contribution by bridging the gap between specialized high-performance tools and comprehensive ecological libraries. As shown in Table @table:metrics, the current landscape is defined by three primary trade-offs: analytical breadth, computational efficiency, and infrastructure overhead.
+While several R packages support diversity analysis, ecodive offers a unique scholarly contribution by bridging the gap between specialized high-performance tools and comprehensive ecological libraries. As shown in Table 1, the current landscape is defined by three primary trade-offs: analytical breadth, computational efficiency, and infrastructure overhead.
 
 | R Package      | Alpha | Beta |      Uni     |      Par     |      Cmp     | Dep  | Citation               |
 |:----------     |   ---:|  ---:|          ---:|          ---:|          ---:|  ---:|:-----------------------|
@@ -65,11 +65,11 @@ While several R packages support diversity analysis, ecodive offers a unique sch
 | `picante`      |    14 |    9 | $\checkmark$ | $-$          | $-$          |   11 | @picante               |
 | `tabula`       |    16 |   12 | $-$          | $-$          | $-$          |    2 | @tabula                |
 | `vegan`        |    17 |   52 | $-$          | $-$          | $\checkmark$ |    7 | @vegan                 |
-Table: Comparison of community diversity metrics and architectural features across 16 R packages. Metrics were identified via an unbiased review of package reference manuals using a large language model to categorize metrics and filter for symmetry and ecological relevance. Alpha: Number of alpha diversity metrics; Beta: Number of beta diversity metrics; Uni: $\checkmark$ indicates UniFrac support; Par: Parallelized calculation of metrics; Cmp: Implementation uses compiled code (e.g. C/C++); Dep: Transitive count of hard R dependencies. Full methodology and LLM prompts are available in `ecodive`'s benchmarking documentation. {@table:metrics}
+Table: Comparison of community diversity metrics and architectural features across 16 R packages. Metrics were identified via an unbiased review of package reference manuals using a large language model to categorize metrics and filter for symmetry and ecological relevance. Alpha: Number of alpha diversity metrics; Beta: Number of beta diversity metrics; Uni: $\checkmark$ indicates UniFrac support; Par: Parallelized calculation of metrics; Cmp: Implementation uses compiled code (e.g. C/C++); Dep: Transitive count of hard R dependencies. Full methodology and LLM prompts are available in `ecodive`'s benchmarking documentation.
 
 ## 1. Analytical Breadth vs. Computational Scalability
 
-The ecological software landscape often requires researchers to choose between a wide variety of metrics and the speed necessary for high-dimensional community data. As shown in Table @table:metrics, packages like `adiv` and `abdiv` offer significant analytical breadth, providing 100 and 65 total metrics, respectively. However, both lack parallelized execution (Par) and compiled backend implementations (Cmp), which limits their utility for large-scale datasets. `ecodive` overcomes this limitation by providing a comprehensive suite of 50 metrics - comparable in scope to the foundational `vegan` package (69 metrics) - while utilizing a parallelized C engine to ensure scalability.
[…]
