+++ paper/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 # Summary
 
-Quantifying the biological complexity of communities - ranging from human gut microbiomes to forest ecosystems - is fundamental to modern ecology. Researchers use diversity metrics to describe the variety within a site (alpha diversity) and the compositional differences between sites (beta diversity). `ecodive` is an R package built to compute these metrics with high efficiency. It provides a comprehensive, unified toolset that allows researchers to analyze large-scale datasets that were previously computationally prohibitive. By integrating optimized C-based algorithms with a parallel processing engine, `ecodive` enables rapid, high-throughput ecological analysis through a lightweight, portable framework.
+Modern ecology relies on quantifying the biological complexity of communities across diverse scales, from microscopic human microbiomes to vast forest ecosystems. Researchers use diversity metrics to describe the variety within a site (alpha diversity) and the compositional differences between sites (beta diversity). `ecodive` is an R package built to compute these metrics with high efficiency. It provides a comprehensive, unified toolset that allows researchers to analyze large-scale datasets that were previously computationally prohibitive. By integrating optimized C-based algorithms with a parallel processing engine, `ecodive` enables rapid, high-throughput ecological analysis through a lightweight, portable framework.
 
 # Statement of Need
 
@@ -45,7 +45,7 @@ Furthermore, the software landscape for ecological metrics is fragmented. A rese
 
 # State of the Field
 
-While several R packages support diversity analysis, ecodive offers a unique scholarly contribution by bridging the gap between specialized high-performance tools and comprehensive ecological libraries. As shown in Table 1, the current landscape is defined by three primary trade-offs: analytical breadth, computational efficiency, and infrastructure overhead.
+While several R packages support diversity analysis, `ecodive` offers a unique scholarly contribution by bridging the gap between specialized high-performance tools and comprehensive ecological libraries. As shown in Table 1, the current landscape is defined by three primary trade-offs: analytical breadth, computational efficiency, and infrastructure overhead.
 
 | R Package      | Alpha | Beta |      Uni     |      Par     |      Cmp     | Dep  | Citation               |
 |:----------     |   ---:|  ---:|          ---:|          ---:|          ---:|  ---:|:-----------------------|
@@ -65,7 +65,7 @@ While several R packages support diversity analysis, ecodive offers a unique sch
 | `picante`      |    14 |    9 | $\checkmark$ | $-$          | $-$          |   11 | @picante               |
 | `tabula`       |    16 |   12 | $-$          | $-$          | $-$          |    2 | @tabula                |
 | `vegan`        |    17 |   52 | $-$          | $-$          | $\checkmark$ |    7 | @vegan                 |
-Table: Comparison of community diversity metrics and architectural features across 16 R packages. Metrics were identified via an unbiased review of package reference manuals using a large language model to categorize metrics and filter for symmetry and ecological relevance. Alpha: Number of alpha diversity metrics; Beta: Number of beta diversity metrics; Uni: $\checkmark$ indicates UniFrac support; Par: Parallelized calculation of metrics; Cmp: Implementation uses compiled code (e.g. C/C++); Dep: Transitive count of hard R dependencies. Full methodology and LLM prompts are available in `ecodive`'s benchmarking documentation.
[…]
