### 2026-05-14 365d78b
better figure integration with text
