### 2026-05-28 46d3e94
bibtex title capitalization
