+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @misc{abdiv,
-  title     = {abdiv: Alpha and Beta Diversity Measures},
+  title     = {abdiv: alpha and beta diversity measures},
   author    = {Bittinger, Kyle},
   year      = {2020},
   journal   = {CRAN: Contributed Packages},
@@ -8,7 +8,7 @@
 }
 
 @article{adiv,
-  title   = {adiv: An r package to analyse biodiversity in ecology},
+  title   = {adiv: An R package to analyse biodiversity in ecology},
   author  = {Sandrine Pavoine},
   year    = {2020},
   journal = {Methods in Ecology and Evolution},
@@ -37,7 +37,7 @@
 }
 
 @article{entropart,
-  title   = {entropart: An R Package to Measure and Partition Diversity},
+  title   = {entropart: an R package to measure and partition diversity},
   author  = {Eric Marcon and Bruno Herault},
   year    = {2015},
   journal = {Journal of Statistical Software},
@@ -48,9 +48,9 @@
 }
 
 @misc{GUniFrac,
-  title     = {GUniFrac: Generalized UniFrac Distances, Distance-Based
-    Multivariate Methods and Feature-Based Univariate Methods for
-    Microbiome Data Analysis},
+  title     = {GUniFrac: generalized UniFrac distances, distance-based
+    multivariate methods and feature-based univariate methods for
+    microbiome data analysis},
   author    = {Jun Chen and Xianyang Zhang and Lu Yang and Lujun Zhang},
   year      = {2023},
   journal   = {CRAN: Contributed Packages},
@@ -59,7 +59,7 @@
 }
 
 @misc{labdsv,
-  title     = {labdsv: Ordination and Multivariate Analysis for Ecology},
+  title     = {labdsv: ordination and multivariate analysis for ecology},
   author    = {Roberts, David W.},
   year      = {2005},
   journal   = {CRAN: Contributed Packages},
@@ -68,7 +68,7 @@
 }
 
 @misc{OmicFlow, 
-  title     = {OmicFlow: Fast and Efficient (Automated) Analysis of Sparse Omics Data},
+  title     = {OmicFlow: fast and efficient (automated) analysis of sparse omics data},
   author    = {Gusinac, Alem and Ederveen, Thomas and Boleij, Annemarie},
   year      = {2025},
   journal   = {CRAN: Contributed Packages},
@@ -77,7 +77,7 @@
 }
 
 @misc{parallelDist,
-  title     = {parallelDist: Parallel Distance Matrix Computation using Multiple Threads},
+  title     = {parallelDist: parallel distance matrix computation using multiple threads},
   author    = {Eckert, Alexander},
   year      = {2017},
   journal   = {CRAN: Contributed Packages},
@@ -86,7 +86,7 @@
 }
 
 @article{philentropy,
-  title   = {Philentropy: Information Theory and Distance Quantification with R},
+  title   = {Philentropy: information theory and distance quantification with R},
   author  = {Hajk-Georg Drost},
   year    = {2018},
   journal = {Journal of Open Source Software},
@@ -131,7 +131,7 @@
 }
 
 @misc{rbiom,
-  title     = {rbiom: Read/Write, Analyze, and Visualize “BIOM” Data},
+  title     = {rbiom: read/write, analyze, and visualize "BIOM" data},
   author    = {Smith, Daniel P.},
   year      = {2020},
   journal   = {CRAN: Contributed Packages},
@@ -141,7 +141,7 @@
 }
 
 @article{tabula,
-  title   = {tabula: An R Package for Analysis, Seriation, and Visualization of Archaeological Count Data},
[…]
