### 2026-05-28 01ba5ab
forced case-sensitivity in bibtex
