+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @misc{abdiv,
-  title     = {abdiv: alpha and beta diversity measures},
+  title     = {{abdiv}: alpha and beta diversity measures},
   author    = {Bittinger, Kyle},
   year      = {2020},
   journal   = {CRAN: Contributed Packages},
@@ -8,7 +8,7 @@
 }
 
 @article{adiv,
-  title   = {adiv: An R package to analyse biodiversity in ecology},
+  title   = {{adiv}: An {R} package to analyse biodiversity in ecology},
   author  = {Sandrine Pavoine},
   year    = {2020},
   journal = {Methods in Ecology and Evolution},
@@ -18,7 +18,7 @@
 }
 
 @article{ampvis2,
-  title     = {ampvis2: an R package to analyse and visualise 16S rRNA amplicon data},
+  title     = {{ampvis2}: an {R} package to analyse and visualise 16S rRNA amplicon data},
   author    = {Andersen, Kasper S. and Kirkegaard, Rasmus H. and Karst, Søren M. and Albertsen, Mads},
   year      = {2018},
   publisher = {openRxiv},
@@ -37,7 +37,7 @@
 }
 
 @article{entropart,
-  title   = {entropart: an R package to measure and partition diversity},
+  title   = {{entropart}: an {R} package to measure and partition diversity},
   author  = {Eric Marcon and Bruno Herault},
   year    = {2015},
   journal = {Journal of Statistical Software},
@@ -59,7 +59,7 @@
 }
 
 @misc{labdsv,
-  title     = {labdsv: ordination and multivariate analysis for ecology},
+  title     = {{labdsv}: ordination and multivariate analysis for ecology},
   author    = {Roberts, David W.},
   year      = {2005},
   journal   = {CRAN: Contributed Packages},
@@ -77,7 +77,7 @@
 }
 
 @misc{parallelDist,
-  title     = {parallelDist: parallel distance matrix computation using multiple threads},
+  title     = {{parallelDist}: parallel distance matrix computation using multiple threads},
   author    = {Eckert, Alexander},
   year      = {2017},
   journal   = {CRAN: Contributed Packages},
@@ -86,7 +86,7 @@
 }
 
 @article{philentropy,
-  title   = {Philentropy: information theory and distance quantification with R},
+  title   = {Philentropy: information theory and distance quantification with {R}},
   author  = {Hajk-Georg Drost},
   year    = {2018},
   journal = {Journal of Open Source Software},
@@ -97,7 +97,7 @@
 }
 
 @article{phyloregion, 
-  title     = {phyloregion: R package for biogeographical regionalization and macroecology},
+  title     = {{phyloregion}: {R} package for biogeographical regionalization and macroecology},
   author    = {Daru, Barnabas H. and Karunarathne, Piyal and Schliep, Klaus},
   year      = {2020},
   journal   = {Methods in Ecology and Evolution},
@@ -108,7 +108,7 @@
 }
 
 @article{phyloseq,
-  title   = {phyloseq: an R package for reproducible interactive analysis 
+  title   = {{phyloseq}: an {R} package for reproducible interactive analysis 
     and graphics of microbiome census data},
   author  = {McMurdie, Paul J. and Holmes, Susan},
   year    = {2013},
@@ -120,7 +120,7 @@
 }
 
 @article{picante,
-  title   = {Picante: R tools for integrating phylogenies and ecology},
+  title   = {Picante: {R} tools for integrating phylogenies and ecology},
   author  = {S.W. Kembel and P.D. Cowan and M.R. Helmus and W.K. Cornwell 
[…]
