### 2026-06-03 e70f237
Added `robust_aitchison()` function. #15
