+++ man/gower.Rd
@@ -28,6 +28,7 @@ class (e.g. \code{phyloseq}). Default: \code{1}}
 \item \code{'percent'}: Relative abundance (sample abundances sum to 1).
 \item \code{'binary'}: Unweighted presence/absence (each count is either 0 or 1).
 \item \code{'clr'}: Centered log ratio.
+\item \code{'rclr'}: Robust centered log ratio.
 }
 
 Default: \code{'none'}.}
@@ -134,6 +135,7 @@ Other Abundance metrics:
 \code{\link{morisita}()},
 \code{\link{motyka}()},
 \code{\link{psym_chisq}()},
+\code{\link{robust_aitchison}()},
 \code{\link{soergel}()},
 \code{\link{squared_chisq}()},
 \code{\link{squared_chord}()},
