+++ reference/robust_aitchison.md
@@ -1,18 +1,15 @@
-# Aitchison distance
+# Robust Aitchison distance
 
-Calculates the Euclidean distance between centered log-ratio (CLR)
-transformed abundances.
+Calculates the pairwise Robust Aitchison distance for compositional
+data. This method is specifically engineered for sparse datasets—such as
+microbiome OTU/ASV tables—by calculating distances based only on
+observed positive abundances, entirely avoiding the need for arbitrary
+pseudo-counts.
 
 ## Usage
 
 ``` r
-aitchison(
-  counts,
-  margin = 1L,
-  pseudocount = NULL,
-  pairs = NULL,
-  cpus = n_cpus()
-)
+robust_aitchison(counts, margin = 1L, pairs = NULL, cpus = n_cpus())
 ```
 
 ## Arguments
@@ -29,11 +26,6 @@ aitchison(
   are columns. Ignored when `counts` is a special object class (e.g.
   `phyloseq`). Default: `1`
 
-- pseudocount:
-
-  Value added to counts to handle zeros when `norm = 'clr'`. Ignored for
-  other normalization methods. See **Pseudocount** section.
-
 - pairs:
 
   Which combinations of samples should distances be calculated for? The
@@ -49,38 +41,29 @@ aitchison(
 
 ## Details
 
-The Aitchison distance is defined as: \$\$\sqrt{\sum\_{i=1}^{n}
-\[(\ln{X_i} - X_L) - (\ln{Y_i} - Y_L)\]^2}\$\$
+The Robust Aitchison distance is defined as: \$\$\sqrt{\sum\_{i=1}^{n}
+(X^\*\_i - Y^\*\_i)^2}\$\$
 
 Where:
 
-- \\X_i\\, \\Y_i\\ : Absolute counts for the \\i\\-th feature.
+- \\X^\*\_i\\, \\Y^\*\_i\\ : The rclr-transformed counts for the
+  \\i\\-th feature. For a given sample \\X\\, \\X^\*\_i = \ln(X_i) -
+  X_L\\ if \\X_i \> 0\\, and \\0\\ otherwise.
 
-- \\X_L\\, \\Y_L\\ : Mean log of abundances. \\X_L =
-  \frac{1}{n}\sum\_{i=1}^{n} \ln{X_i}\\.
+- \\X_L\\, \\Y_L\\ : Mean log of strictly positive abundances. \\X_L =
+  \frac{1}{\|P_X\|}\sum\_{j \in P_X} \ln{X_j}\\, where \\P_X\\ is the
+  set of indices where \\X \> 0\\.
 
-- \\n\\ : The number of features.
+- \\\|P_X\|\\, \\\|P_Y\|\\ : The number of strictly positive features in
+  the respective samples.
+
+- \\n\\ : The total number of features.
 
 Base R Equivalent:
 
-    x <- log((x + pseudocount) / exp(mean(log(x + pseudocount))))
-    y <- log((y + pseudocount) / exp(mean(log(y + pseudocount))))
+    x <- ifelse(x > 0, log(x) - mean(log(x[x > 0])), 0)
+    y <- ifelse(y > 0, log(y) - mean(log(y[y > 0])), 0)
     sqrt(sum((x-y)^2)) # Euclidean distance
-
-## Pseudocount
-
-Zeros are undefined in the Aitchison (CLR) transformation. If
-`pseudocount` is `NULL` (the default) and zeros are detected, the
-function uses half the minimum non-zero value (`min(x[x>0]) / 2`) and
-issues a warning.
-
-To suppress the warning, provide an explicit value (e.g., `1`).
-
-**Why this matters:** The choice of pseudocount is not neutral; it acts
-as a weighting factor that can significantly distort downstream results,
[…]

[…]
