+++ R/beta_div.r
@@ -1040,9 +1040,9 @@ psym_chisq <- function (counts, margin = 1L, pairs = NULL, cpus = n_cpus()) {
 #' Robust Aitchison distance
 #' 
 #' Calculates the pairwise Robust Aitchison distance for compositional data. 
-#' This method is specifically engineered for sparse datasets—such as microbiome 
-#' OTU/ASV tables—by calculating distances based only on observed positive abundances, 
-#' entirely avoiding the need for arbitrary pseudo-counts.
+#' This method is specifically engineered for sparse datasets - such as microbiome 
+#' OTU/ASV tables - by calculating distances based only on observed positive 
+#' abundances, avoiding the need for pseudo-counts.
 #' 
 #' @inherit documentation
 #' 
