+++ vignettes/articles/bdiv_guide.Rmd
@@ -38,7 +38,7 @@ This decision tree presents a series of questions to help a researcher systemati
 
 * **YES:** You must account for the relative nature of your data (the "constant sum" constraint).
 
-    * The most statistically rigorous approach is to use a compositional metric. -> Your primary choice should be Robust Aitchison distance. While the standard Aitchison metric requires adding arbitrary pseudo-counts to handle the zeroes inherent in microbiome data, the robust version natively calculates distances using only observed positive abundances (**[Martino et al. (2019)](https://doi.org/10.1128/mSystems.00016-19)**). **[Gloor et al. (2017)](https://doi.org/10.3389/fmicb.2017.02224)** argue this compositional approach is essential to avoid the "false positives" common in standard ecological metrics, and the robust adaptation makes it ideal for highly sparse sequencing datasets.
+    * The most statistically rigorous approach is to use a compositional metric. -> Your primary choice should be <u>Robust Aitchison</u> distance. While the standard Aitchison metric requires adding arbitrary pseudo-counts to handle the zeroes inherent in microbiome data, the robust version natively calculates distances using only observed positive abundances (**[Martino et al. (2019)](https://doi.org/10.1128/mSystems.00016-19)**). **[Gloor et al. (2017)](https://doi.org/10.3389/fmicb.2017.02224)** argue this compositional approach is essential to avoid the "false positives" common in standard ecological metrics, and the robust adaptation makes it ideal for highly sparse sequencing datasets.
     
     * If you choose to use a non-compositional metric, select one that is robust to sparsity. -> Good secondary choices include <u>Hellinger</u> or <u>Jensen-Shannon Divergence (JSD)</u>. In a rigorous benchmarking of 13 metrics on high-dimensional data, **[Chen et al. (2021)](https://doi.org/10.1371/journal.pone.0246893)** demonstrated that Hellinger and JSD captured compositional changes in low-abundance elements more efficiently than other standard metrics. Similarly, **[Cordier et al. (2024)](https://doi.org/10.1093/ismeco/ycaf230)** found Hellinger to be one of the top performers in handling noisy, sparse count data.
 
