+++ conda/meta.yaml
@@ -1,63 +1,0 @@
-{% set version = '1.0.0' %}
-{% set posix = 'm2-' if win else '' %}
-
-package:
-  name: r-ecodive
-  version: {{ version|replace("-", "_") }}
-
-source:
-  url:
-    - {{ cran_mirror }}/src/contrib/ecodive_{{ version }}.tar.gz
-    - {{ cran_mirror }}/src/contrib/Archive/ecodive/ecodive_{{ version }}.tar.gz
-  sha256: d5ab2cfbe40968a37455401e28bf96fce2cef45bc79ddcbc8a6a75fc31aa76d5
-
-build:
-  number: 0
-  rpaths:
-    - lib/R/lib/
-    - lib/
-  missing_dso_whitelist:
-    - '*/R.dll'        # [win]
-    - '*/Rblas.dll'    # [win]
-    - '*/Rlapack.dll'  # [win]
-
-requirements:
-  build:
-    - cross-r-base {{ r_base }}    # [build_platform != target_platform]
-    - {{ stdlib("c") }}            # [not win]
-    - {{ compiler('c') }}          # [not win]
-    - {{ compiler('cxx') }}        # [not win]
-    - {{ stdlib("m2w64_c") }}      # [win]
-    - {{ compiler('m2w64_c') }}    # [win]
-    - {{ compiler('m2w64_cxx') }}  # [win]
-    - {{ posix }}filesystem        # [win]
-    - {{ posix }}make
-    - {{ posix }}sed               # [win]
-    - {{ posix }}coreutils         # [win]
-    - {{ posix }}zip               # [win]
-  host:
-    - r-base
-  run:
-    - r-base
-
-test:
-  commands:
-    - $R -e "library('ecodive')"           # [not win]
-    - "\"%R%\" -e \"library('ecodive')\""  # [win]
-
-about:
-  home: https://cmmr.github.io/ecodive, https://github.com/cmmr/ecodive
-  summary: Computes alpha and beta diversity metrics using concurrent 'C' threads. Metrics include
-    'UniFrac', Faith's phylogenetic diversity, Bray-Curtis dissimilarity, Shannon diversity
-    index, and many others. Also parses newick trees into 'phylo' objects and rarefies
-    feature tables.
-  license: MIT
-  license_family: MIT
-  license_file:
-    - '{{ environ["PREFIX"] }}/lib/R/share/licenses/MIT'
-    - LICENSE
-
-extra:
-  recipe-maintainers:
-    - conda-forge/r
-    - dansmith01
