### 2026-06-04 05cee33
custom dictionary for CRAN's aspell check
