+++ .aspell/defaults.R
@@ -1,0 +1,1 @@
+description <- list(dictionaries = c("en_stats", "ecodive"))
