+++ DESCRIPTION
@@ -27,7 +27,6 @@ URL: https://cmmr.github.io/ecodive/, https://github.com/cmmr/ecodive
 BugReports: https://github.com/cmmr/ecodive/issues
 License: MIT + file LICENSE
 Encoding: UTF-8
-Language: en-US
 LazyData: true
 Depends: R (>= 3.6.0)
 Roxygen: list(markdown = TRUE)
